refactor(client): :recycle: Move most processing down to the Department component
