+++ client/src/components/UI/MapFrance/MapFrance.tsx
@@ -1,46 +1,11 @@
-import { cls } from "lib/classname";
-import { useEffect, useState } from "react";
+import React from "react";
 import Department from "./Department";
-import styles from "./MapFrance.module.scss";
 
 interface Props {
-  colors: Record<string, string[]>;
   onSelectDep: (dep: string) => void;
-  selectable: string[];
 }
 
-const normalizeColors = (colors: Record<string, string[]>) => {
-  const normalizedColors: Record<string, string> = {};
-  for (const [color, depts] of Object.entries(colors)) {
-    for (const dept of depts) {
-      normalizedColors[dept] = color;
-    }
-  }
-  return normalizedColors;
-};
-
-const MapFrance = ({ onSelectDep, colors, selectable }: Props) => {
-  const [activeDep, setActiveDep] = useState("");
-  const [colorsPerDep, setColorsPerDep] = useState<Record<string, string>>(normalizeColors(colors));
-
-  useEffect(() => {
-    onSelectDep(activeDep);
-  }, [activeDep, onSelectDep]);
-
-  useEffect(() => {
-    setColorsPerDep(normalizeColors(colors));
-  }, [colors]);
-
-  const getDepProps = (dep: string) => {
-    const isSelectable = selectable.includes(dep);
-    return {
-      id: `dpt-${dep}`,
-      onClick: () => (isSelectable ? setActiveDep((d) => (d === dep ? "" : dep)) : {}),
-      fill: colorsPerDep[dep] || undefined,
-      className: cls(styles.dep, isSelectable && styles.selectable, dep === activeDep && styles.selected),
-    };
-  };
-
+const MapFrance: React.FC<Props> = ({ onSelectDep }) => {
   return (
     <svg width="578px" height="544px" viewBox="0 0 578 544" style={{ maxWidth: "100%" }}>
       <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
@@ -95,436 +60,436 @@ const MapFrance = ({ onSelectDep, colors, selectable }: Props) => {
           </g>
           <g id="carte" transform="translate(12.000000, 2.000000)" fill="#ADADF9" stroke="#E3E3FD" strokeWidth="0.4">
             <Department
+              dep="974"
               d="M194.995,496.96 L192.664,497.005 L191.817,497.42 L190.852,498.163 L190.674,498.326 L190.006,498.326 L188.595,498.86 L187.987,499.39 L186.042,499.672 L185.864,500.696 L185.329,502.834 C185.329,502.834 184.585,503.331 184.364,503.413 C184.143,503.495 182.909,504.155 182.909,504.155 L182.686,506.426 L183.22,507.78 L184.675,509.503 L185.373,510.29 L185.937,510.171 L185.596,510.334 L185.64,513.423 L187.095,515.235 L188.595,516.631 L191.238,518.193 L191.416,518.089 L191.208,518.237 L194.341,520.123 L196.761,521 L199.359,521.445 L201.438,522.186 L202.759,522.439 C202.759,522.439 204.796,522.024 205.061,522.024 C205.326,522.024 207.436,521.608 207.436,521.608 L209.604,520.91 L211.059,519.589 L210.925,517.243 L211.193,515.772 L211.772,514.941 L212.737,513.664 L213.227,512.06 L213.093,510.498 L212.069,509.919 L210.747,509.315 L210.139,509.271 L209.159,507.875 C208.938,507.71 207.747,506.388 207.659,506.185 C207.571,505.981 206.783,504.207 206.783,504.207 L206.516,501.445 L205.016,500.005 L203.249,499.026 C203.249,499.026 201.836,498.207 201.571,498.166 C201.306,498.125 198.971,497.631 198.795,497.631 C198.618,497.631 196.894,497.753 196.85,497.794 L196.627,497.705 L194.995,496.96 Z"
-              {...getDepProps("974")}
             ></Department>
             <Department
+              dep="973"
               d="M128.604,494.792 C128.527,494.797 128.47,494.82 128.443,494.846 L126.962,496.719 L125.677,498.2 L124.196,499.679 L124.41,503.64 L124.303,507.298 L125.588,509.082 L126.855,511.349 L127.586,512.745 L126.855,513.638 L125.999,516.011 L126.106,519.282 L125.053,521.156 L123.358,522.548 L123.876,523.137 L125.161,523.922 L126.321,524.618 L128.016,524.118 L129.818,523.026 L133.529,522.935 L134.921,523.916 L136.402,524.023 L137.883,524.326 L140.007,522.631 L142.452,518.385 L144.254,514.727 L146.056,512.05 L147.966,509.587 L149.019,507.695 L148.18,506.012 L147.538,503.639 L146.164,503.139 L144.594,502.074 L143.291,500.879 L141.596,499.79 L140.436,499.394 L137.046,497.025 L134.816,496.218 L132.692,495.932 C132.692,495.932 131.741,495.541 131.211,495.343 C130.746,495.172 129.14,494.757 128.604,494.792 Z"
-              {...getDepProps("973")}
             ></Department>
             <Department
+              dep="972"
               d="M74.279,495.505 C74.279,495.505 70.667,495.593 70.225,495.757 C69.784,495.922 68.013,497.153 68.013,497.153 L67.746,497.222 L66.692,498.795 L67.048,500.845 L70.137,503.889 L69.602,504.72 L69.602,506.368 L70.923,508.179 L75.17,511.965 L78.957,511.965 L79.313,511.965 L80.367,511.876 L80.545,511.876 L80.367,512.618 L80.367,513.197 L81.614,513.524 L82.134,513.846 C82.134,513.846 81.96,515.247 82.402,515.329 L81.08,515.404 L79.224,514.84 L77.992,514.587 L76.403,515.24 L75.512,516.307 L75.081,517.536 L75.422,518.783 L76.313,520.014 L77.456,520.663 C77.456,520.663 78.603,520.34 79.045,520.174 C79.486,520.012 79.847,519.273 79.847,519.273 C79.847,519.273 81.159,519.192 81.778,519.11 C82.396,519.027 83.025,519.852 83.025,519.852 L85.935,520.248 L87.257,520.991 L88.044,520.917 L88.044,520.086 L88.757,519.69 L89.47,520.086 L89.114,521.08 L88.312,521.822 L88.668,523.292 L89.722,523.871 L90.969,523.798 L92.201,522.804 L92.557,521.496 L93.789,520.175 L92.987,518.529 L92.467,517.044 L91.843,515.069 L91.219,513.673 C91.219,513.673 89.719,512.194 89.719,511.861 C89.719,511.528 88.308,510.792 88.308,510.792 L88.13,509.476 C88.13,509.476 88.49,509.077 88.932,508.912 C89.373,508.748 88.665,508.008 88.665,508.008 L87.789,508.17 L86.453,508.496 L86.111,508.081 L85.576,507.502 L85.665,506.76 L86.63,506.943 L87.432,506.037 L86.288,505.132 L85.397,504.879 L85.129,504.392 L85.129,503.816 L86.109,503.4 L86.807,503.976 L88.039,504.464 L88.663,503.722 L88.663,502.91 L89.197,502.257 L88.752,501.752 L87.787,501.354 L87.163,501.916 L84.698,502.169 L83.896,502.748 L82.842,502.421 L81.61,500.446 L78.774,497.634 L77.185,496.986 L74.279,495.505 Z"
-              {...getDepProps("972")}
             ></Department>
             <Department
+              dep="971"
               d="M18.939,497.65 L17.439,498.004 L16.058,499.222 L16.118,501.953 L17.187,502.132 L17.247,503.824 L16.193,504.636 L16,505.979 L16,506.216 L15.065,507.196 L15.005,507.374 L12.941,507.256 L12.689,506.261 L11.189,505.405 L8.516,504.009 L5.576,504.009 L4.21,506.1 L4.27,509.119 L4.641,510.348 L4.701,510.576 L4.894,512.907 L5.77,517.214 L6.646,519.485 C6.646,519.485 7.264,520.07 7.389,520.36 C7.514,520.651 8.205,521.58 8.205,521.58 L9.571,521.758 L9.957,521.046 L10.016,521.12 L11.382,521.12 L13.446,519.367 L14.753,517.569 L14.693,513.381 L13.936,511.168 L14.188,509.356 L15.614,509.079 L15.629,509.079 L16.06,509.6 L17.248,509.882 L18.184,510.236 L18.273,510.191 L18.377,510.535 L20.189,510.639 L23.307,510.06 L25.059,508.785 L28.118,508.67 L30.731,507.794 L29.112,506.872 L28.919,505.994 L26.855,504.658 L22.935,504.596 L21.747,502.562 L21.747,500.291 L20.93,498.539 L18.939,497.65 Z"
-              {...getDepProps("971")}
             ></Department>
             <Department
+              dep="976"
               d="M260.945,522.599 C260.69,522.599 260.492,522.621 260.369,522.872 C260.22,523.178 259.939,522.855 259.765,522.802 C259.431,522.698 259.183,522.864 259.047,523.172 C258.992,523.298 259.024,523.409 258.987,523.535 C258.938,523.706 258.761,523.668 258.753,523.508 C258.725,522.944 258.789,522.34 258.128,522.175 C257.439,522.004 257.158,522.327 257.295,522.986 C257.318,523.092 257.34,523.215 257.326,523.329 C257.234,523.328 257.136,523.348 257.077,523.426 C256.978,523.559 257.013,523.748 256.902,523.877 C256.789,524.008 256.457,524.176 256.284,524.187 C255.967,524.208 255.984,523.686 255.852,523.499 C255.724,523.311 255.389,523.197 255.189,523.35 C255.072,523.441 255.117,523.702 254.904,523.621 C254.796,523.392 254.821,523.059 254.744,522.809 C254.666,522.554 254.294,522.454 254.048,522.454 C253.778,522.454 253.588,523.164 253.267,523.342 C253.184,523.397 253.096,522.843 253.222,522.728 C253.47,522.502 253.339,522.352 253.36,521.999 C253.382,521.639 253.737,520.585 252.96,520.798 C252.745,520.858 252.708,521.089 252.499,521.152 C252.268,521.222 251.951,521.311 251.715,521.214 C251.517,521.132 251.589,520.887 251.279,520.93 C250.939,520.978 250.934,520.808 251.208,520.652 C251.429,520.522 251.646,520.41 251.896,520.346 C252.197,520.272 252.275,520.034 252.489,519.839 C252.821,519.539 252.938,518.827 252.781,518.411 C252.713,518.226 252.665,517.975 252.436,517.95 C252.274,517.929 251.697,518.539 251.697,518.012 C251.697,517.672 251.623,517.71 251.378,517.533 C251.152,517.367 251.082,517.061 251.001,516.808 C250.932,516.593 250.959,516.246 250.756,516.108 C250.595,516.004 250.26,515.999 250.318,515.722 C250.662,515.678 251.047,515.618 251.394,515.623 C251.533,515.626 251.652,515.704 251.791,515.715 C251.938,515.725 252.052,515.658 252.191,515.633 C252.328,515.606 252.658,515.834 252.716,515.946 C252.787,516.089 252.734,516.258 252.747,516.402 C252.767,516.623 252.819,516.826 252.83,517.048 C252.844,517.337 253.067,517.471 253.218,517.692 C253.399,517.96 253.68,517.921 253.941,518.045 C254.21,518.174 254.443,518.281 254.732,518.365 C255.057,518.462 255.285,518.733 255.635,518.761 C255.829,518.775 256.164,518.704 256.309,518.561 C256.445,518.423 256.769,518.112 256.766,517.957 C256.763,517.816 256.777,517.678 256.762,517.539 C256.808,517.435 256.913,517.531 256.886,517.359 C256.836,517.042 256.506,516.853 256.41,516.558 C256.378,516.46 256.44,516.353 256.319,516.291 C256.214,516.232 256.104,516.251 256.078,516.1 C256.021,515.769 256.086,515.752 255.816,515.556 C255.607,515.406 255.671,515.118 255.524,514.925 C255.449,514.828 255.334,514.85 255.239,514.787 C255.075,514.677 254.951,514.516 254.797,514.394 C254.364,514.053 253.748,513.814 253.638,513.205 C253.597,512.975 253.584,512.734 253.277,512.733 C252.984,512.732 252.639,512.539 252.47,512.296 C252.377,512.165 252.078,512.047 252.16,511.832 C252.199,511.728 252.364,511.747 252.442,511.696 C252.559,511.623 252.557,511.54 252.602,511.423 C252.779,510.96 253.902,511.158 253.942,510.902 C253.979,510.666 253.769,510.645 253.592,510.616 C253.347,510.571 253.165,510.412 252.897,510.386 C252.603,510.358 252.754,510.03 252.859,509.872 C253.053,509.587 253.054,509.226 253.054,508.893 C253.054,508.395 252.323,508.084 252.637,507.478 C252.778,507.207 253.105,507.426 253.258,507.224 C253.571,506.809 252.785,506.978 252.648,506.964 C252.391,506.943 252.3,506.599 252.324,506.389 C252.352,506.143 252.568,506.046 252.685,505.844 C252.763,505.71 252.952,505.274 252.894,505.125 C252.784,504.844 252.666,504.699 252.4,504.91 C251.881,505.32 251.557,504.389 251.08,504.64 C250.835,504.773 250.586,504.72 250.455,504.452 C250.388,504.306 250.281,504.336 250.172,504.25 C250.072,504.173 249.978,504.034 249.889,503.938 C249.726,503.769 249.556,503.613 249.312,503.586 C249.108,503.566 248.807,503.863 248.627,503.774 C248.52,503.717 248.493,503.565 248.412,503.488 C248.305,503.383 248.167,503.378 248.038,503.319 C247.701,503.17 248.192,502.774 248.283,502.671 C248.562,502.344 248.974,502.007 249.121,501.634 C249.383,500.971 248.184,501.279 247.933,501.125 C247.712,500.992 248.511,500.375 248.558,500.171 C248.613,499.934 248.483,499.669 248.6,499.442 C248.742,499.171 248.86,499.307 249.092,499.363 C249.332,499.417 249.678,499.432 249.804,499.185 C249.873,499.05 250.065,498.706 250.188,498.623 C250.338,498.518 250.563,498.237 250.374,498.048 C250.08,497.754 250.228,497.808 250.602,497.757 C250.769,497.736 251.15,497.774 251.176,497.543 C251.198,497.352 251.176,496.891 251.341,496.761 C251.521,496.619 251.654,496.454 251.834,496.319 C251.917,496.257 252.04,496.206 252.096,496.11 C252.179,495.968 252.028,495.756 252.27,495.727 C252.528,495.703 252.544,495.812 252.741,495.558 C252.815,495.464 253.264,494.999 253.247,495.37 C253.239,495.552 252.788,495.699 253.033,495.884 C253.299,496.083 252.979,496.115 252.831,496.155 C252.407,496.259 251.415,496.634 252.173,497.051 C252.409,497.18 252.657,497.239 252.912,497.29 C253.33,497.377 252.602,497.639 253.078,497.728 C253.229,497.757 253.274,497.72 253.357,497.845 C253.398,497.907 253.456,498.046 253.52,498.085 C253.761,498.231 253.666,498.439 253.788,498.664 C253.896,498.864 254.206,499.017 254.399,499.129 C254.546,499.216 254.705,499.297 254.881,499.297 C255.03,499.297 255.246,499.245 255.333,499.413 C255.42,499.582 254.568,499.854 254.84,500.285 C254.982,500.508 255.263,500.62 255.369,500.88 C255.457,501.096 255.31,501.55 255.632,501.509 C255.752,501.493 255.905,501.424 256.019,501.505 C256.112,501.569 256.125,501.63 256.257,501.636 C256.568,501.651 256.567,501.44 256.832,501.72 C257.424,502.349 257.74,501.389 258.19,501.215 C258.481,501.104 258.51,501.156 258.696,501.365 C258.805,501.487 258.975,501.803 259.189,501.716 C259.473,501.592 259.687,501.758 259.94,501.709 C260.236,501.647 260.348,501.688 260.598,501.849 C260.782,501.967 261.111,501.812 261.328,501.913 C261.418,501.955 261.57,501.951 261.634,502.04 C261.676,502.099 261.625,502.181 261.751,502.18 C261.908,502.179 261.943,501.963 262.027,502.228 C262.057,502.323 262.123,502.338 262.062,502.418 C261.978,502.526 262.231,502.71 262.283,502.779 C262.452,503.004 262.746,502.99 263.006,503.008 C263.211,503.029 263.412,503.321 263.504,503.487 C263.582,503.623 263.696,503.616 263.783,503.717 C263.904,503.864 264.092,504.071 264.156,504.252 C264.323,504.721 263.3,505.206 263.693,505.781 C263.836,505.99 264.014,506.073 264.039,506.364 C264.075,506.784 263.664,506.426 263.501,506.524 C263.145,506.736 262.42,506.991 262.343,507.457 C262.266,507.916 262.044,507.881 261.755,508.15 C261.587,508.301 261.172,508.495 261.176,508.778 C261.18,509.12 261.162,509.197 260.873,509.398 C260.629,509.567 260.447,509.548 260.33,509.846 C260.286,509.956 260.368,510.024 260.435,510.088 C260.57,510.232 260.432,510.24 260.397,510.377 C260.326,510.648 260.194,510.912 260.376,511.18 C260.457,511.299 260.474,511.395 260.519,511.527 C260.553,511.626 260.473,511.718 260.545,511.8 C260.619,511.885 260.695,511.999 260.782,512.071 C260.865,512.135 261.064,512.113 261.114,512.214 C261.262,512.514 261.116,512.645 261.521,512.7 C261.663,512.721 262.181,512.929 261.805,512.982 C261.718,513.261 261.921,513.628 262.228,513.628 C262.321,513.629 262.423,513.581 262.515,513.614 C262.654,513.664 262.385,513.832 262.37,513.873 C262.263,514.171 262.366,514.535 262.112,514.769 C261.885,514.978 261.554,515.118 261.286,515.273 C261.078,515.392 260.812,515.627 260.687,515.838 C260.555,516.065 260.708,516.687 260.729,516.942 C260.739,517.066 261.212,517.612 260.661,517.38 C260.192,517.192 259.969,517.355 259.771,517.737 C259.647,517.979 259.322,517.936 259.108,518.113 C258.919,518.267 259.018,518.779 259.114,518.967 C259.294,519.321 259.182,519.374 258.797,519.358 C258.424,519.337 258.327,519.901 258.434,520.17 C258.56,520.487 258.902,520.566 258.928,520.941 C258.95,521.28 259.283,521.417 259.561,521.544 C259.742,521.626 259.915,521.683 260.077,521.802 C260.182,521.882 260.27,522.024 260.409,522.038 C260.57,522.059 260.752,521.989 260.893,522.099 C261.127,522.278 260.955,522.349 260.991,522.578 L260.945,522.599 Z"
-              {...getDepProps("976")}
             ></Department>
             <Department
+              dep="976"
               d="M268.594,508.362 C268.566,508.424 268.447,508.404 268.391,508.422 C268.237,508.47 268.098,508.485 267.987,508.613 C267.941,508.667 267.927,508.719 267.905,508.787 C267.867,508.891 267.813,508.996 267.767,509.098 C267.71,509.222 267.756,509.404 267.636,509.494 C267.579,509.539 267.516,509.536 267.443,509.544 C267.332,509.557 267.324,509.569 267.252,509.483 C267.192,509.414 267.162,509.386 267.155,509.295 C267.149,509.209 267.155,509.119 267.155,509.031 C267.155,508.948 267.164,508.901 267.099,508.843 C267.037,508.789 266.962,508.752 266.896,508.701 C266.836,508.655 266.757,508.603 266.712,508.543 C266.659,508.473 266.642,508.376 266.602,508.3 C266.502,508.113 266.456,507.955 266.261,507.875 C266.177,507.84 266.119,507.773 266.039,507.732 C265.93,507.67 265.829,507.643 265.733,507.564 C265.567,507.423 265.4,507.286 265.433,507.043 C265.603,506.961 265.666,507.108 265.799,507.176 C265.925,507.24 266.045,507.306 266.159,507.384 C266.304,507.488 266.44,507.508 266.464,507.304 C266.473,507.225 266.45,507.158 266.443,507.08 C266.44,507.042 266.454,506.993 266.444,506.957 C266.435,506.922 266.402,506.902 266.387,506.873 C266.362,506.823 266.335,506.728 266.341,506.674 C266.352,506.583 266.46,506.578 266.525,506.528 C266.587,506.479 266.623,506.406 266.671,506.349 C266.725,506.287 266.8,506.236 266.87,506.192 C266.942,506.147 267.025,506.142 267.092,506.087 C267.164,506.027 267.212,505.983 267.28,505.902 C267.322,505.854 267.38,505.817 267.422,505.763 C267.485,505.692 267.516,505.604 267.565,505.524 C267.61,505.449 267.689,505.328 267.767,505.279 C267.897,505.198 267.955,505.318 268.011,505.412 C268.038,505.458 268.076,505.499 268.103,505.543 C268.162,505.647 268.172,505.688 268.291,505.681 C268.438,505.679 268.512,505.752 268.539,505.884 C268.565,506.016 268.56,506.138 268.575,506.275 C268.582,506.334 268.61,506.379 268.641,506.428 C268.689,506.507 268.685,506.545 268.696,506.636 C268.709,506.745 268.733,506.872 268.794,506.963 C268.834,507.025 268.894,507.031 268.936,507.086 C268.975,507.137 268.967,507.216 268.958,507.281 C268.739,507.274 268.636,507.499 268.637,507.689 C268.637,507.917 268.729,508.117 268.694,508.355 C268.658,508.373 268.618,508.383 268.576,508.376 L268.594,508.362 Z"
-              {...getDepProps("976")}
             ></Department>
             <Department
-              {...getDepProps("95")}
+              dep="95"
               points="286.814 102.337 285.296 104.213 284.15 108.389 282.879 110.537 287.483 112.595 290.993 111.96 295.347 114.744 300.007 114.955 303.305 116.953 304.158 119.737 304.128 119.798 304.491 119.737 307.88 117.922 312.964 117.589 315.748 116.318 317.533 115.047 318.108 111.325 316.867 110.084 312.721 107.633 308.728 105.545 306.669 106.483 304.399 107.058 302.88 105.908 299.859 104.032 297.408 105.908 293.804 106.302 288.695 105.908 287.546 104.032"
             ></Department>
             <Department
-              {...getDepProps("94")}
+              dep="94"
               points="36.085 31.82 33.496 33.911 30.907 34.309 30.708 35.703 30.21 36.997 31.504 36.997 32.201 36.399 31.902 35.403 35.885 35.204 36.881 35.602 36.881 37.693 35.885 39.784 30.11 38.987 29.214 38.191 26.526 40.083 23.937 40.78 20.253 40.581 19.158 44.464 20.154 46.157 18.561 50.837 20.154 53.824 22.942 52.43 29.613 56.114 41.163 56.114 48.034 59.301 47.934 58.604 52.813 46.357 52.813 40.781 49.726 39.984 48.93 38.49 47.636 37.893 44.848 34.806 39.969 32.018"
             ></Department>
             <Department
-              {...getDepProps("93")}
+              dep="93"
               points="49.926 4.836 44.051 9.018 34.891 13.2 18.163 14.295 18.661 16.585 20.055 16.884 21.349 18.875 20.055 21.663 18.363 21.962 19.558 24.85 27.623 24.75 30.212 28.733 30.71 34.309 33.498 33.911 36.087 31.82 39.97 32.019 44.849 34.807 47.637 37.894 48.931 38.491 49.727 39.985 52.814 40.782 52.814 31.721 49.13 9.814"
             ></Department>
             <Department
-              {...getDepProps("92")}
+              dep="92"
               points="18.163 14.295 7.01 20.27 5.815 20.469 0.338 32.418 3.823 41.877 9.4 48.15 9.4 48.847 18.063 54.921 20.154 53.826 18.561 50.839 20.154 46.159 19.158 44.466 20.253 40.583 20.054 40.583 13.98 37.496 9.001 36.102 8.304 33.912 8.204 32.02 11.191 28.635 14.477 28.436 15.971 26.445 19.556 24.852 18.361 21.964 20.053 21.665 21.347 18.877 20.053 16.886 18.659 16.587"
             ></Department>
             <Department
-              {...getDepProps("91")}
+              dep="91"
               points="304.881 128.21 303.186 128.967 301.34 129.693 300.977 131.751 298.198 133.052 297.835 135.079 299.137 137.318 297.261 139.89 294.507 139.89 295.596 141.555 294.295 143.038 293.841 146.034 294.749 146.216 295.112 148.667 295.505 149.242 295.9 154.326 301.957 153.782 304.408 151.512 306.496 153.206 311.58 153.569 312.306 154.658 312.306 150.845 316.543 148.061 315.272 146.396 315.819 141.373 316.548 140.284 315.64 135.261 317.304 133.96 316.971 131.6 314.883 130.632 311.373 130.632 309.346 129.512 307.866 130.268"
             ></Department>
             <Department
-              {...getDepProps("90")}
+              dep="90"
               fillRule="nonzero"
               points="475.494 176.265 476.065 177.413 479.091 177.595 482.511 180.258 483.055 181.559 482.873 183.829 481.936 185.554 482.298 187.824 484.96 187.43 486.473 193.542 483.964 193.937 483.239 194.784 471.587 189.519 472.616 188.46 473.073 185.434 473.073 183.558 472.257 180.683 472.257 178.262 473.739 177.173"
             ></Department>
             <Department
-              {...getDepProps("89")}
+              dep="89"
               points="347.7 148.485 346.005 149.786 338.621 149.423 335.232 151.118 333.9 153.962 335.414 155.657 333.143 158.32 331.417 160.378 334.837 163.616 335.775 166.642 338.254 169.275 338.254 172.694 333.322 176.84 335.011 178.747 334.653 181.561 331.99 183.468 328.208 183.468 328.782 185.556 331.233 188.945 331.803 191.971 332.378 194.059 331.258 194.392 333.654 194.936 335.44 194.936 336.378 193.453 337.589 193.453 338.95 194.663 338.678 196.176 339.919 196.993 341.674 196.993 344.125 198.627 345.215 198.083 346.304 199.021 347.394 198.749 349.27 197.539 351.055 198.084 352.266 197.812 352.266 194.423 352.931 194.544 353.358 196.299 355.539 197.661 355.539 199.719 358.656 199.84 362.741 203.653 365.308 203.774 365.162 202.564 366.252 200.93 367.22 202.14 366.408 203.502 366.802 204.743 368.284 203.775 370.064 203.775 369.918 206.498 371.553 207.588 372.793 207.043 375.814 204.985 375.632 204.591 373.847 203.502 373.726 201.595 375.632 200.657 376.448 199.568 375.904 198.63 375.904 196.451 377.383 194.272 379.562 189.642 379.958 187.614 381.318 187.07 381.465 186.676 380.77 186.132 380.77 184.498 382.954 182.864 383.498 180.141 382.56 178.537 381.047 178.537 380.652 178.113 380.652 175.813 382.554 174.451 382.408 173.089 382.136 171.728 381.167 173.816 379.836 173.634 378.716 171.546 374.934 173.452 367.372 173.058 366.434 170.97 364.346 168.156 363.985 164.737 360.962 161.135 359.056 162.467 355.636 159.835 356.211 154.721 351.309 149.607 349.039 149.607"
             ></Department>
             <Department
-              {...getDepProps("88")}
+              dep="88"
               points="482.812 137.439 480.874 137.863 477.908 138.983 476.426 140.829 474.762 141.949 471.584 143.431 470.676 143.25 468.442 142.675 466.231 142.493 465.472 140.072 463.993 140.647 463.806 143.25 462.691 142.13 460.665 143.613 458.638 144.34 456.217 144.158 454.552 144.733 452.706 143.613 451.586 144.521 448.803 144.521 447.501 146.942 441.389 147.305 440.087 145.641 438.06 145.278 438.247 144.34 439.549 143.432 438.429 143.039 438.247 140.073 435.463 139.891 433.829 140.406 432.679 141.949 429.714 142.857 429.714 141.768 428.8 141.768 428.8 143.039 425.078 143.795 425.078 145.46 421.749 145.883 420.478 147.154 421.783 149.03 424.804 149.241 427.467 152.812 429.738 152.812 430.282 155.293 432.761 155.656 433.306 157.351 431.617 158.501 430.497 163.615 435.793 166.248 435.793 170.212 437.851 169.456 439.58 170.817 440.881 170.03 440.153 168.91 441.091 168.335 442.392 169.848 444.299 168.516 444.873 166.64 448.443 166.065 449.412 166.822 449.225 168.91 451.676 170.786 453.195 170.605 454.884 169.273 458.848 169.273 461.698 172.481 462.636 172.481 464.151 171.543 464.365 170.423 465.879 169.666 467.573 170.786 469.263 172.693 475.225 175.84 477.585 173.813 477.766 166.429 481.004 163.615 482.88 159.258 482.306 156.413 484.577 151.874 486.665 147.153 483.061 146.215 482.305 141.101 483.061 137.682"
             ></Department>
             <Department
-              {...getDepProps("87")}
+              dep="87"
               points="271.291 258.363 269.597 260.057 264.846 259.694 264.301 259.573 260.7 260.239 258.612 261.964 258.612 264.234 254.466 264.416 252.015 267.26 250.502 268.38 251.803 269.893 251.621 275.007 250.683 276.702 252.196 278.396 254.829 278.607 255.404 281.241 255.585 282.935 252.196 283.691 250.502 284.265 250.865 288.986 248.595 290.501 246.689 291.075 245.751 293.739 244.238 293.921 243.602 296.796 248.02 296.945 248.988 298.823 248.201 300.911 249.926 303.182 251.227 302.819 252.559 300.917 258.611 301.487 259.579 305.27 263.725 305.843 264.663 307.145 261.818 308.477 263.15 309.415 267.145 309.99 267.447 310.746 268.507 308.6 270.777 309.451 273.258 306.362 274.68 306.546 275.921 304.485 279.613 304.912 281.065 303.459 282.917 303.459 284.369 300.797 286.603 300.584 288.056 298.134 290.118 298.738 291.568 298.738 292.539 297.075 291.568 296.471 291.141 294.022 292.594 294.022 292.594 292.996 290.115 291.33 287.452 290.121 287.846 287.642 287.034 287.642 286.005 289.094 282.707 289.279 280.859 289.094 280.859 287.217 283.13 285.368 281.677 283.522 278.197 284.549 278.802 276.738 277.168 275.075 275.504 270.961 273.053 270.356 272.629 268.086 274.899 263.577 274.838 260.793 274.324 261.398"
             ></Department>
             <Department
-              {...getDepProps("86")}
+              dep="86"
               points="222.117 214.666 218.516 218.448 217.608 220.022 217.911 224.682 219.545 224.47 219.756 226.528 220.392 229.826 221.421 232.096 220.18 233.518 220.785 234.547 219.968 236 219.968 236.605 221.421 238.27 221.421 239.299 220.786 240.933 218.728 244.019 220.786 244.836 221.421 246.289 220.574 248.528 220.392 249.557 219.151 251.615 219.151 253.068 219.968 253.28 219.968 257.608 221.421 258.637 220.786 260.059 220.997 261.3 222.631 263.146 223.478 261.905 223.478 260.876 225.112 260.059 226.353 260.876 226.353 263.962 225.112 265.415 224.083 267.866 225.505 270.136 228.41 270.953 227.775 272.83 225.082 273.283 227.563 276.308 230.861 276.097 233.736 275.068 236.641 276.732 237.67 275.916 237.459 273.223 239.093 271.983 240.546 274.463 241.756 275.704 245.266 274.252 246.719 272.617 249.987 272.617 251.682 273.465 251.803 269.894 250.502 268.381 252.015 267.261 254.466 264.417 258.612 264.235 258.612 261.965 260.7 260.24 265.814 259.302 266.177 256.457 264.119 255.337 262.969 251.373 259.943 250.979 258.067 249.103 254.466 246.258 255.223 243.988 255.223 240.387 251.804 236.998 251.441 234.335 248.203 230.946 247.083 226.407 245.751 225.832 244.238 223.744 242.725 224.682 243.119 226.77 238.186 227.92 232.527 227.92 232.709 225.65 232.709 222.049 228.169 220.717 228.169 218.447 224.568 217.691 223.812 214.665"
             ></Department>
             <Department
-              {...getDepProps("85")}
+              dep="85"
               points="173.487 218.418 172.398 220.446 169.281 220.446 170.522 221.717 169.584 224.834 166.77 225.772 165.681 224.985 166.165 221.868 165.378 220.143 163.653 220.143 162.412 221.535 163.017 225.923 164.439 227.95 163.017 229.524 160.384 229.04 156.481 228.102 155.392 225.137 152.88 224.834 149.611 223.412 148.824 221.536 145.253 219.206 139.625 226.469 139.443 231.008 145.283 236.667 145.101 238.362 146.796 238.362 150.397 249.165 154.18 251.041 157.962 254.824 162.32 254.824 164.014 258.607 168.19 258.607 170.066 261.451 174.242 263.539 174.424 260.876 175.483 261.875 181.263 258.365 183.926 258.183 185.046 261.209 188.647 259.696 191.673 261.966 194.124 260.816 196.212 260.241 197.15 259.121 199.782 257.427 197.906 255.52 196.393 256.67 196.211 255.157 197.543 252.676 196.211 250.406 197.331 248.893 196.756 243.597 194.88 240.783 196.212 239.058 193.186 235.669 194.306 233.218 189.767 229.435 189.767 227.165 187.891 224.502 188.254 224.351 185.682 222.172 180.84 222.172 179.266 221.234 177.087 220.931 174.575 218.571"
             ></Department>
             <Department
-              {...getDepProps("84")}
+              dep="84"
               points="414.093 365.396 411.43 365.607 409.372 368.816 409.916 372.206 413.124 372.6 412.579 374.118 410.098 374.3 407.286 377.148 406.53 376.21 407.075 372.427 405.955 371.096 400.841 371.854 399.843 373.88 400.387 374.184 403.595 379.509 403.595 383.807 409.224 389.405 409.224 391.83 407.045 393.067 407.196 393.129 414.035 396.031 416.939 399.269 418.332 401.172 422.084 402.867 426.532 402.716 433.886 405.775 438.002 406.293 443.148 405.111 445.262 403.445 445.533 402.023 441.595 397.577 437.298 397.577 437.298 396.032 438.841 394.306 438.841 392.429 435.422 390.736 435.089 387.982 436.965 387.131 436.965 384.74 434.907 384.377 434.756 381.808 434.731 381.623 433.002 381.502 430.157 379.414 429.4 376.963 424.103 376.567 420.137 376.207 419.741 373.936 421.073 371.088 418.622 373.178 414.839 372.782 414.083 371.451 416.716 367.88"
             ></Department>
             <Department
-              {...getDepProps("83")}
+              dep="83"
               points="481.634 395.232 478.759 395.353 477.403 396.745 472.257 396.562 467.809 399.795 464.752 397.737 459.94 399.281 459.091 401.006 455.672 403.578 449.498 399.463 444.505 401.101 444.264 402.433 444.356 402.371 447.444 404.432 443.331 407.186 441.637 410.604 443.699 412.297 443.512 415.901 446.783 418.991 445.905 419.681 442.148 419.499 442.148 420.861 443.842 421.708 443.842 422.223 442.298 423.619 443.176 424.642 444.537 424.642 445.895 426.004 446.081 427.032 444.352 428.057 442.995 429.602 442.808 433.357 443.328 434.047 446.718 435.566 447.656 439.343 449.744 439.739 451.647 438.407 455.04 436.319 460.911 436.892 460.724 438.407 458.822 439.345 463.362 439.558 462.242 438.412 461.846 435.961 464.297 434.267 467.142 435.205 468.291 435.568 469.229 436.714 470.561 435.776 470.923 433.296 472.438 431.994 476.405 431.994 477.551 430.266 480.183 431.026 483.209 429.719 483.209 424.786 479.245 424.968 482.266 423.092 483.784 421.004 484.178 417.982 489.657 417.223 492.714 413.803 490.596 411.622 490.596 410.412 489.571 409.385 490.929 408.203 490.596 406.298 488.358 405.449 487.177 405.449 485.118 403.39 484.756 399.786 482.546 398.759 480.307 398.61 479.46 396.55"
             ></Department>
             <Department
-              {...getDepProps("82")}
+              dep="82"
               points="252.559 368.634 251.227 372.961 253.315 375.231 252.377 378.835 251.227 380.167 252.014 382.438 248.413 383.194 245.569 383.953 246.325 386.586 244.903 387.01 245.751 387.797 248.504 387.797 249.866 387.131 251.076 387.797 248.837 388.826 248.837 390.551 246.779 392.094 246.779 394.303 250.562 394.488 251.742 395.699 251.742 398.092 253.134 399.451 251.924 401.328 252.438 402.357 254.829 401.328 259.792 401.328 261.184 400.48 262.878 400.48 264.754 402.716 269.051 401.506 270.231 399.629 268.537 398.752 269.717 398.271 270.08 396.543 271.956 396.728 272.985 397.908 273.832 396.549 276.586 395.52 277.827 396.094 278.795 394.49 277.1 392.615 280.338 392.615 281.365 390.738 283.605 388.676 281.365 388.676 282.062 385.776 287.873 385.081 290.112 383.72 292.836 382.691 293.713 381.814 292.654 379.09 294.197 376.003 291.468 375.821 291.135 371.55 287.025 373.426 284.267 372.761 282.577 374.822 279.672 374.459 279.158 377.88 277.766 377.88 276.918 376 276.404 375.152 274.861 375.298 271.956 378.385 268.355 378.238 267.84 376 267.84 374.3 267.507 373.938 266.297 375.298 264.754 375.298 263.059 373.422 261.516 373.422 260.487 371.363 261.849 369.82 261.698 368.822 259.277 368.913 259.58 369.579 255.404 370.335"
             ></Department>
             <Department
-              {...getDepProps("81")}
+              dep="81"
               points="301.582 380.436 298.312 380.623 296.099 382.832 295.408 381.289 293.865 382.166 293.714 381.804 292.837 382.685 290.113 383.71 287.874 385.067 282.062 385.762 281.367 388.666 283.607 388.666 281.367 390.725 280.339 392.602 277.101 392.602 278.796 394.477 277.828 396.081 278.131 396.202 278.978 398.745 281.036 401.165 281.55 401.499 281.701 403.891 283.427 406.643 284.788 408.188 284.788 410.063 283.427 410.583 284.455 412.126 288.903 414.002 289.751 415.212 291.99 417.941 294.557 418.632 296.619 418.273 296.619 420.694 296.557 421.931 297.677 421.785 297.406 423.116 301.158 423.116 304.637 424.206 304.908 420.726 306.512 420.726 312.382 421.788 318.011 421.788 322.282 420.456 323.372 416.977 320.434 414.287 321.221 411.625 322.825 409.229 325.488 411.08 332.448 408.686 333.537 406.54 329.997 404.269 328.454 403.909 327.425 405.267 325.367 405.784 322.096 404.092 318.677 399.645 318.677 397.768 316.8 394.319 316.987 392.776 315.441 388.84 312.539 387.479 310.477 383.876 304.824 382.848"
             ></Department>
             <Department
-              {...getDepProps("80")}
+              dep="80"
               points="281.217 39.212 280.701 45.446 284.878 49.228 284.878 51.135 279.764 48.109 273.621 55.674 275.255 56.37 277.343 56.189 278.281 58.065 286.598 66.205 287.173 69.594 289.049 73.013 288.413 73.8 290.804 73.8 291.2 74.375 291.2 76.463 292.348 77.25 294.827 76.282 298.306 76.282 300.035 77.614 306.541 77.039 311.353 78.401 314.041 81.094 317.49 81.457 320.94 82.607 325.172 78.583 330.164 77.251 332.857 77.614 336.851 77.22 336.488 75.132 335.55 71.289 335.154 68.989 336.881 67.051 336.881 64.57 339.36 61.877 338.906 59.214 337.454 58.82 335.764 57.126 331.224 57.126 328.376 58.639 328.376 57.307 327.649 55.975 325.921 56.731 323.831 57.881 324.98 56.187 324.795 54.674 322.316 52.949 322.316 54.856 320.622 54.856 316.658 53.524 314.966 52.192 313.238 53.342 312.693 54.855 311.937 55.037 310.999 54.099 310.999 51.436 313.812 49.56 312.692 47.654 310.421 48.622 309.275 47.654 304.552 48.229 302.103 49.167 299.622 48.229 299.622 46.353 296.232 44.84 295.294 42.57 293.962 43.115 290.359 40.301 288.665 39.726 287.335 41.421 285.064 41.996 284.126 39.908"
             ></Department>
             <Department
-              {...getDepProps("79")}
+              dep="79"
               points="215.308 218.267 210.194 218.448 205.473 219.386 200.359 219.779 200.359 222.412 197.726 224.137 191.855 222.805 187.891 224.499 189.767 227.162 189.767 229.432 194.306 233.215 193.186 235.666 196.212 239.055 194.88 240.78 196.756 243.594 197.331 248.89 196.211 250.403 197.543 252.673 196.211 255.154 196.393 256.667 197.906 255.517 199.782 257.424 197.15 259.118 196.212 260.238 194.124 260.813 191.673 261.963 191.491 261.812 191.582 264.596 193.64 266.835 193.64 268.076 196.727 270.346 200.631 270.558 203.294 274.038 207.622 273.644 211.132 276.095 213.795 277.33 214.218 279.392 215.701 280.814 217.486 279.816 217.093 278.788 218.515 277.124 218.515 275.701 220.996 275.064 224.295 273.429 227.775 272.825 228.41 270.948 225.505 270.131 224.083 267.861 225.112 265.41 226.353 263.957 226.353 260.871 225.112 260.054 223.478 260.871 223.478 261.9 222.631 263.141 220.997 261.295 220.786 260.054 221.421 258.632 219.968 257.603 219.968 253.275 219.151 253.063 219.151 251.61 220.392 249.552 220.574 248.523 221.421 246.284 220.786 244.831 218.728 244.014 220.786 240.928 221.421 239.294 221.421 238.265 219.968 236.6 219.968 236 220.785 234.547 220.18 233.518 221.421 232.096 220.392 229.826 219.756 226.528 219.545 224.47 217.911 224.682 217.608 220.022 217.215 220.718 215.883 219.598"
             ></Department>
             <Department
-              {...getDepProps("78")}
+              dep="78"
               points="282.881 110.538 282.82 110.659 277.524 112.172 276.768 113.11 277.736 114.623 277.736 116.136 279.43 116.499 278.462 117.467 278.553 118.314 278.674 118.224 280.55 120.1 281.124 122.37 282.639 123.883 281.881 125.971 281.881 127.847 283.212 129.36 281.881 131.63 283.212 134.868 285.845 136.744 286.241 138.832 288.511 139.195 289.056 143.371 290.785 145.459 293.842 146.034 294.296 143.038 295.597 141.555 294.508 139.89 297.262 139.89 299.138 137.318 297.836 135.079 298.199 133.052 300.983 131.751 301.341 129.693 303.192 128.967 304.882 128.21 305.214 128.422 305.214 128.21 303.525 126.304 302.463 123.429 304.157 119.737 303.304 116.953 300.006 114.955 295.346 114.744 290.992 111.96 287.482 112.595"
             ></Department>
             <Department
-              {...getDepProps("77")}
+              dep="77"
               fillRule="nonzero"
               points="318.104 111.438 320.466 109.9 322.542 111.229 323.492 111.958 325.186 111.776 326.336 110.263 328.968 111.776 332.206 110.444 334.082 111.019 335.958 109.506 337.108 108.931 339.197 110.447 339.772 114.623 344.856 119.525 346.55 120.1 347.519 122.37 350.604 123.083 350.151 123.702 350.363 124.822 348.85 125.79 348.456 128.423 349.97 129.361 350.726 132.206 349.788 133.9 350.362 135.232 353.176 135.05 353.179 136.561 350.546 137.681 350.364 139.951 349.033 141.676 348.276 145.459 347.833 148.417 346.005 149.786 338.621 149.423 335.232 151.118 333.9 153.962 335.414 155.657 331.418 160.378 326.729 162.938 326.337 161.528 324.248 161.346 323.674 162.859 321.797 163.222 316.684 163.04 314.595 164.372 313 162.833 315.928 160.767 315.745 157.563 313.475 156.413 312.294 154.658 312.294 150.845 316.533 148.061 315.26 146.396 315.833 141.417 316.521 140.333 315.624 135.261 317.29 133.96 316.925 131.388 318.406 127.688 318.422 123.188 317.288 116.56"
             ></Department>
             <Department
-              {...getDepProps("76")}
+              dep="76"
               fillRule="nonzero"
               points="228.137 81.396 232.677 74.587 240.521 70.417 249.896 67.022 264.27 63.421 272.38 57.187 273.621 55.674 275.255 56.37 277.343 56.189 278.281 58.065 286.603 66.205 287.178 69.594 288.979 73 287.178 75.285 286.5 77.125 287.854 77.188 287.166 79 286.421 82.85 287.541 84.726 287.562 87.896 289.354 88 287.359 91.746 286.844 93.32 283.85 92.987 278.812 90.396 274.468 90.536 272.168 92.563 270.928 96.104 266.056 96.649 261.583 98.021 262.979 99.062 262.394 100.734 260.229 100.667 259.021 98.188 257.643 98.131 255.583 96.521 258.416 95.562 253.833 92.292 249.23 92.562 248.146 90.708 245.538 90.958 243.667 87.708 242.028 88.084 239.849 90.414 236.208 90.833 235.875 90.542 230.982 89.536 227.199 86.691"
             ></Department>
             <Department
-              {...getDepProps("75")}
+              dep="75"
               points="27.622 24.751 19.557 24.851 15.972 26.444 14.478 28.435 11.192 28.634 8.205 32.019 8.305 33.911 9.002 36.101 13.981 37.495 20.055 40.582 23.938 40.781 26.527 40.084 29.215 38.192 30.111 38.988 35.886 39.785 36.882 37.694 36.882 35.603 35.886 35.205 31.903 35.404 32.202 36.4 31.505 36.998 30.211 36.998 30.709 35.704 30.908 34.31 30.709 34.31 30.211 28.734"
             ></Department>
             <Department
+              dep="75-92-93-94"
               d="M308.108648,130.182479 L307.530827,130.485068 L304.827665,128.589765 L304.827665,128.372277 L303.087443,126.414879 L302,123.463336 L303.709018,119.734825 L304.0819,119.67273 L307.562031,117.808318 L307.568375,117.934276 L312.52433,117.609864 L315.382574,116.304932 L317.215782,115 L316.967401,116.553312 L318.11694,123.389071 L318.11694,126.035656 C318.179187,126.09712 318.241698,126.158936 318.304473,126.221105 L318.304473,127.961014 L316.782052,131.782511 L316.813256,132 L314.669261,131.005543 L311.065252,131.005543 L308.983664,129.856005 L308.113709,130.290982 L308.108648,130.182479 Z"
-              {...getDepProps("75-92-93-94")}
             ></Department>
             <Department
-              {...getDepProps("74")}
+              dep="74"
               points="471.347 253.884 467.019 254.64 462.843 258.029 461.724 256.335 459.636 256.516 457.729 260.692 457.946 262.387 460.004 264.081 456.222 266.562 453.771 268.832 449.595 268.832 447.689 270.345 445.842 270.95 444.814 272.827 442.968 271.798 442.544 277.124 443.57 279 443.631 281.027 445.205 282.267 445.205 287.621 448.927 288.225 450.349 290.917 453.652 291.311 454.256 290.1 455.92 290.1 458.795 293.188 459.823 294.217 463.122 293.579 463.934 291.945 464.75 288.646 466.597 287.193 468.049 282.896 469.897 281.657 471.348 282.051 471.744 283.291 470.716 284.532 472.562 286.803 475.653 286.803 477.53 289.889 477.101 290.917 478.952 289.678 480.405 288.042 481.676 288.193 481.767 284.894 488.213 282.23 488.969 280.355 488.575 276.179 484.429 271.851 483.098 272.607 483.098 270.88 483.098 268.066 479.496 266.341 479.309 264.828 481.401 262.558 481.401 259.926 477.979 256.325 477.796 253.874 471.351 253.874"
             ></Department>
             <Department
+              dep="73"
               d="M443.143,280.635 L442.75,286.173 L441.509,287.201 L441.509,293.375 L439.03,293.799 L438.634,297.309 L436.364,298.944 L436.334,298.918 L436.576,299.98 L438.846,302.429 L438.846,304.092 L441.085,308.208 L445.836,309.233 L447.471,310.473 L448.111,309.836 L447.259,308.202 L448.499,305.51 L449.952,305.327 L452.403,308.596 L453.426,307.173 L454.879,307.173 L458.782,311.077 L459,313.74 L457.154,314.769 L456.301,318.667 L457.965,320.332 L457.965,321.968 L457.541,323.63 L458.782,322.394 L460.84,323.63 L461.93,325.112 L465.349,324.75 L466.469,325.869 L467.043,328.532 L470.434,328.136 L470.83,325.292 L472.524,324.534 L476.156,324.716 L476.094,324.597 L481.752,322.326 L483.84,323.658 L485.928,323.658 L486.115,321.387 L488.565,320.055 L489.503,318.936 L494.436,317.032 L495.011,313.823 L494.042,312.309 L496.705,307.77 L494.254,306.832 L493.497,304.169 L488.383,301.145 C488.383,301.145 488.694,295.346 488.202,294.305 C488.182,294.284 488.153,294.231 488.14,294.214 C488.135,294.208 488.119,294.19 488.115,294.184 C488.111,294.184 488.087,294.184 488.085,294.184 C488.085,294.178 488.084,294.16 488.085,294.159 C488.081,294.159 488.057,294.159 488.055,294.159 C488.052,294.158 488.033,294.159 488.029,294.159 C488.021,294.159 488.002,294.159 488,294.159 C487.243,294.347 484.429,294.555 484.429,294.555 L481.584,291.317 L481.675,288.202 L480.404,288.051 L478.953,289.687 L477.102,290.926 L477.529,289.898 L475.652,286.812 L472.563,286.812 L470.717,284.541 L471.745,283.3 L471.349,282.06 L469.896,281.666 L468.05,282.905 L466.598,287.203 L464.751,288.655 L463.935,291.955 L463.123,293.588 L459.824,294.226 L458.796,293.197 L455.921,290.111 L454.257,290.111 L453.653,291.32 L450.353,290.926 L448.931,288.234 L445.209,287.63 L445.209,282.276 L443.143,280.635 Z"
-              {...getDepProps("73")}
             ></Department>
             <Department
-              {...getDepProps("72")}
+              dep="72"
               points="235.371 143.189 230.257 143.37 225.143 148.091 222.48 147.94 218.939 149.09 217.85 150.966 217.366 155.021 217.699 157.502 214.885 159.862 214.098 161.405 214.734 162.827 213.947 164.703 214.25 165.49 211.133 165.793 210.679 167.185 212.071 170.635 211.92 171.573 210.679 172.662 208.016 172.813 207.683 173.751 208.318 174.689 208.318 180.317 207.713 181.739 209.407 183.131 209.407 185.007 212.221 185.158 219.575 187.337 227.533 192.179 230.105 193.178 231.376 190.636 235.159 193.299 237.247 193.299 236.097 189.516 238.397 191.029 239.698 189.123 245.175 187.61 244.237 185.34 245.569 183.646 248.413 182.526 251.046 179.106 251.046 175.505 252.953 175.505 253.709 172.873 253.891 168.909 252.015 167.184 253.528 164.552 255.767 161.707 253.134 159.831 250.683 159.437 248.02 155.473 247.263 155.473 247.082 157.349 246.9 156.048 242.936 156.048 241.03 153.203 237.822 152.053 236.884 145.244"
             ></Department>
             <Department
-              {...getDepProps("71")}
+              dep="71"
               points="379.686 216.754 379.686 216.936 376.841 216.936 374.541 218.297 373.18 220.204 374.269 221.838 375.207 222.382 374.541 224.016 373.573 224.56 372.511 226.043 373.722 226.86 373.576 228.767 373.847 230.794 375.632 231.883 375.632 232.7 373.726 233.79 373.726 235 375.087 236.211 374.936 236.907 371.97 236.756 371.699 237.724 370.338 237.996 368.977 239.751 367.221 239.751 365.865 241.113 364.08 241.384 363.142 239.084 361.932 238.933 358.966 240.779 362.476 247.376 362.476 250.221 363.414 251.371 366.622 251.371 367.741 252.884 370.762 252.884 372.093 254.76 371.911 262.144 367.947 265.17 367.854 265.382 368.125 265.352 368.7 265.534 368.882 268.56 371.726 269.135 372.089 270.467 373.606 270.467 375.877 269.135 381.929 270.073 383.079 271.223 384.592 269.71 386.68 269.71 387.8 263.658 388.556 263.083 390.433 263.083 392.738 264.596 394.427 263.083 395.546 264.596 397.274 262.901 399.335 262.719 400.303 265.745 401.032 269.528 402.364 269.709 403.51 267.046 407.081 253.065 408.412 250.614 410.5 250.403 412.588 252.128 414.102 251.734 415.978 250.403 417.884 250.796 419.004 253.247 420.093 253.671 425.056 253.066 426.962 251.553 426.206 250.403 423.935 249.646 423.723 247.013 425.629 245.682 426.386 242.474 424.693 239.448 423.547 237.935 424.121 237.36 424.121 235.484 422.603 234.516 422.209 233.003 426.563 232.459 426.959 230.946 425.628 230.946 424.508 229.614 422.42 229.614 420.695 226.769 419.177 226.588 419.363 224.318 417.724 224.136 415.545 225.618 413.366 226.163 411.884 225.225 409.98 225.104 409.433 225.891 407.677 226.163 403.743 227.404 398.023 229.431 395.839 227.676 393.268 227.131 393.151 224.831 390.7 223.045 387.307 220.473 383.495 220.594 379.834 216.811"
             ></Department>
             <Department
-              {...getDepProps("70")}
+              dep="70"
               points="448.438 166.066 444.872 166.641 444.298 168.517 442.391 169.849 441.09 168.336 440.152 168.911 440.88 170.031 439.004 171.181 439.579 172.694 437.49 173.45 437.49 175.901 434.828 175.901 434.641 177.414 432.008 177.989 432.195 180.259 433.884 180.441 432.946 181.561 432.372 185.344 430.677 185.344 427.832 186.494 424.807 185.344 422.174 186.494 422.174 188.401 424.236 188.764 425.749 191.79 425.961 193.303 423.51 196.147 422.364 196.51 421.604 197.448 423.695 198.416 424.053 201.442 425.747 201.623 425.959 205.406 426.716 206.163 426.807 206.768 428.259 207.162 429.896 208.524 432.345 208.524 433.283 207.586 434.947 207.646 436.673 207.707 440.485 204.59 441.575 204.59 442.785 203.652 446.72 203.773 449.715 201.322 451.895 200.928 452.711 198.901 454.466 198.356 456.222 195.36 458.673 193.453 461.244 193.06 463.151 194.422 466.691 193.998 466.691 192.091 468.114 191.274 469.233 189.731 471.412 189.731 472.623 188.46 473.076 185.434 473.076 183.558 472.26 180.683 472.26 178.262 473.742 177.173 475.497 176.265 475.316 175.902 469.258 172.694 467.569 170.787 465.874 169.667 464.361 170.424 464.149 171.544 462.631 172.482 461.693 172.482 458.843 169.274 454.879 169.274 453.19 170.606 451.671 170.787 449.22 168.911 449.407 166.823"
             ></Department>
             <Department
-              {...getDepProps("69")}
+              dep="69"
               points="399.325 262.72 397.263 262.902 395.537 264.597 394.418 263.084 392.724 264.597 390.424 263.084 388.547 263.084 387.791 263.659 387.396 265.838 388.42 266.655 388.42 268.501 389.661 270.559 388.632 271.8 384.734 271.8 384.521 275.491 383.282 276.73 384.094 279.818 384.094 281.665 385.759 283.299 385.759 285.569 388.209 287.84 388.209 290.503 386.999 292.35 388.421 293.201 388.421 294.624 387.184 295.649 387.184 297.313 392.538 302.25 398.923 303.274 399.528 305.125 397.466 307.576 399.13 308.6 401.192 307.788 403.401 308.12 407.365 306.149 406.73 304.913 404.067 302.25 408.179 300.798 413.533 299.56 417.648 294.236 414.955 291.965 415.106 289.997 411.261 288.877 408.782 289.693 408.386 286.394 406.115 284.332 404.663 284.121 400.975 281.881 401.579 280.431 401.579 274.044 402.607 272.621 402.968 268.204 402.33 269.717 400.998 269.536 400.271 265.753"
             ></Department>
             <Department
-              {...getDepProps("68")}
+              dep="68"
               points="489.199 147.698 486.349 147.849 484.569 151.873 482.298 156.412 482.872 159.257 480.996 163.614 477.758 166.426 477.577 173.81 475.217 175.837 475.308 175.898 476.069 177.411 479.09 177.593 482.51 180.256 483.054 181.557 482.872 183.827 481.934 185.552 482.296 187.823 484.958 187.429 485.533 189.517 486.471 193.541 488.71 193.178 488.316 195.236 489.647 196.386 496.633 196.204 500.237 193.36 500.419 189.184 502.325 186.733 499.844 183.888 498.543 180.862 500.056 178.804 500.056 174.053 500.994 171.783 500.994 168 502.686 165.549 500.811 162.916 500.625 157.469 497.72 156.198 496.056 151.386 492.333 151.568 490.854 149.722"
             ></Department>
             <Department
-              {...getDepProps("67")}
+              dep="67"
               points="503.846 104.849 500.244 105.848 498.58 108.753 498.58 111.598 497.066 112.93 495.735 112.93 493.284 111.205 491.378 112.537 489.108 112.537 487.232 110.661 483.628 110.086 481.54 109.148 480.783 106.304 479.089 108.18 478.151 112.537 475.67 113.294 475.67 115.745 478.151 116.895 480.028 118.227 479.271 119.921 480.996 121.041 484.022 118.771 489.288 121.797 487.049 125.973 487.236 127.305 488.749 128.818 487.599 132.782 483.822 136.565 481.734 136.383 483.06 137.684 482.304 141.103 483.06 146.217 486.664 147.155 486.361 147.851 489.206 147.7 490.87 149.727 492.349 151.573 496.077 151.391 497.741 156.203 500.641 157.474 500.616 156.869 505.549 147.216 504.974 141.739 507.245 134.355 507.82 127.909 512.717 124.338 512.717 122.068 514.623 119.587 516.136 119.587 517.831 117.893 517.468 114.685 519.162 110.145 521.795 109.57 519.162 107.482 514.44 106.937 510.26 104.849 507.415 106.543 505.902 104.849 503.846 104.849"
             ></Department>
             <Department
-              {...getDepProps("66")}
+              dep="66"
               points="330.33 448.945 326.215 451.154 326.547 453.575 308.603 453.575 308.754 456.813 310.63 459.567 302.943 463.319 302.127 462.835 295.802 463.289 295.016 464.893 291.808 465.71 289.662 467.586 283.792 468.948 284.124 470.945 286.969 473.608 292.628 475.121 292.81 478.511 295.835 481.173 298.105 480.78 301.343 476.815 305.307 476.054 311.541 478.147 316.837 482.681 318.35 480.78 319.681 480.78 321.013 481.718 322.132 481.173 322.314 478.511 328.003 477.179 329.879 474.729 332.724 473.791 336.688 473.791 339.169 476.423 342.195 476.635 342.195 473.614 340.676 471.525 338.014 470.376 337.59 453.855 335.111 452.708"
             ></Department>
             <Department
-              {...getDepProps("65")}
+              dep="65"
               points="212.887 411.09 211.011 412.089 211.344 412.119 214.552 417.414 212.464 419.321 213.977 421.591 216.247 424.985 214.371 427.431 211.345 434.243 206.231 438.597 207.563 441.26 206.413 442.016 203.599 441.447 202.812 447.676 201.299 448.825 200.996 452.726 201.48 452.455 204.688 454.361 208.471 457.206 208.834 459.477 211.86 461.925 214.341 461.925 220.575 459.262 223.208 462.288 226.809 463.256 228.14 460.986 229.835 461.743 233.497 461.985 233.255 451.817 235.162 451.817 236.856 452.665 238.066 451.484 237.884 449.608 240.275 448.212 239.428 444.644 238.399 443.764 236.341 444.459 237.37 442.734 236.855 440.525 233.769 438.286 233.951 436.743 235.676 433.838 237.885 432.99 237.885 431.78 239.247 429.718 240.155 428.447 236.705 426.662 232.075 426.662 231.379 425.268 228.988 425.268 228.292 423.724 226.083 423.724 225.387 424.417 222.815 424.417 222.664 422.873 220.606 421.515 221.302 420.821 221.635 419.127 221.12 418.612 220.091 415.857 217.852 415.525 215.643 414.314 215.794 411.077 212.887 411.077"
             ></Department>
             <Department
-              {...getDepProps("64")}
+              dep="64"
               points="206.048 411.423 203.324 412.967 197.847 412.785 197.332 411.937 193.731 413.148 190.978 413.995 188.92 412.452 186.711 413.148 186.015 412.269 183.291 412.269 181.566 413.148 177.117 412.961 174.726 414.836 170.429 414.504 169.582 415.72 168.553 415.357 169.914 413.995 167.523 412.119 164.618 414.691 159.655 415.023 154.208 412.295 153.421 413.691 149.064 418.988 145.675 420.319 143.224 420.677 143.224 422.765 145.494 424.859 148.883 425.04 149.065 427.491 151.728 427.703 152.484 425.974 156.085 427.492 158.355 428.067 158.9 430.337 157.599 431.487 157.599 435.052 154.936 436.384 154.754 438.078 156.449 439.984 159.475 440.922 160.05 438.077 161.744 436.201 161.562 438.652 162.894 440.553 166.283 440.553 167.796 442.615 172.335 443.372 176.693 446.035 183.865 446.035 184.258 449.999 189.16 453.781 191.067 456.051 193.155 454.932 195.031 454.538 195.969 455.476 197.694 454.538 200.992 452.723 201.295 448.824 202.808 447.675 203.595 441.446 206.409 442.015 207.559 441.259 206.227 438.596 211.341 434.242 214.367 427.43 216.243 424.979 213.973 421.59 212.46 419.32 214.548 417.413 211.34 412.117 206.226 411.755"
             ></Department>
             <Department
-              {...getDepProps("63")}
+              dep="63"
               fillRule="nonzero"
               points="320.222 274.311 321.765 273.614 322.976 270.709 324.156 270.376 324.188 268.891 326.188 268.703 326.391 269.562 328.968 266.775 331.871 266.769 332.386 267.798 331.025 269.674 333.052 271.914 333.415 273.275 338.196 276.029 343.922 276.891 345.731 276.695 348.303 277.209 350.512 275.848 352.238 276.695 352.57 279.087 354.734 279.609 357.714 279.451 358.562 281.508 361.203 282.516 361.135 283.328 364.094 285.781 362.466 287.473 361.892 291.65 364.162 294.675 366.22 299.577 371.153 302.814 372.781 309.031 369.245 312.438 369.82 314.708 371.219 315.234 370.545 315.915 368.154 315.219 366.469 313.531 365.422 315.5 362.495 315.4 361.141 314.203 360.062 316.469 358.229 315.734 357.031 313.516 355.475 313.495 354.114 312.314 351.984 313.141 349.665 313.346 348.303 312.466 347.453 312.922 347.109 312.141 345.731 312.829 345.688 313.969 343.673 314.191 341.797 315.736 338.377 316.25 337.591 317.369 336.834 317.279 336.138 318.641 333.415 318.458 330.843 314.523 329.814 314.373 327.756 313.346 326.516 314.469 323.489 314.525 321.945 311.287 316.135 309.593 316.5 308.312 315.926 305.083 314.776 303.751 314.776 301.089 316.484 298.266 316.108 295.611 311.748 291.255 311.385 289.562 317.047 286.375 317.625 283.734 319.859 282 319.133 278.395 317.609 276.469 317.375 272.156"
             ></Department>
             <Department
-              {...getDepProps("62")}
+              dep="62"
               points="300.069 4.503 289.599 6.5 281.278 12.916 281.278 38.457 281.217 39.213 284.153 39.909 285.091 41.997 287.362 41.422 288.693 39.727 290.383 40.302 293.987 43.116 295.319 42.571 296.257 44.841 299.651 46.354 299.651 48.23 302.13 49.168 304.581 48.23 309.302 47.655 310.451 48.623 312.721 47.655 313.841 49.561 311.029 51.437 311.029 54.1 311.967 55.038 312.723 54.856 313.268 53.343 314.997 52.193 316.687 53.525 320.656 54.857 322.345 54.857 322.345 52.95 324.826 54.675 325.008 56.188 323.862 57.882 325.951 56.732 327.676 55.976 328.405 57.308 328.405 58.64 331.25 57.127 335.79 57.127 335.971 57.309 337.091 55.161 336.516 54.192 334.791 53.829 332.885 53.829 331.737 53.435 333.641 52.285 335.37 52.467 337.095 52.285 337.307 49.198 338.456 48.442 338.638 46.717 336.52 45.355 334.041 45.173 333.467 44.779 335.01 43.629 335.373 42.479 334.041 41.541 332.103 39.211 332.316 38.061 334.796 36.911 335.191 35.579 333.465 34.792 332.496 32.311 329.046 31.917 325.597 30.979 325.202 27.136 327.681 25.593 326.743 23.656 324.805 23.656 323.474 25.774 317.512 25.411 312.7 24.231 310.038 21.356 310.038 19.056 312.126 18.088 310.398 16.756 306.193 16.574 303.678 10.219"
             ></Department>
             <Department
-              {...getDepProps("61")}
+              dep="61"
               fillRule="nonzero"
               points="194.859 128.594 192.794 127.273 196.606 126.123 199.42 122.855 203.323 123.944 206.742 121.584 208.8 120.827 210.979 123.006 214.58 122.37 217.697 124.095 221.601 122.854 225.172 120.191 227.532 117.528 229.106 117.225 229.562 119.172 230.801 118.95 230.952 117.528 234.553 116.923 235.794 117.679 239.638 116.832 238.759 119.344 240.272 121.25 245.568 121.432 246.656 123.984 248.594 125.971 251.227 126.909 252.014 129.754 250.683 131.63 254.071 133.899 254.253 137.137 258.247 140.526 258.609 147.469 256.734 149.786 253.89 150.361 251.045 152.055 251.439 153.568 252.74 155.293 252.719 159.703 250.682 159.441 248.016 155.5 247.262 155.477 247.094 157 246.922 156.078 242.935 156.052 241.031 153.203 237.812 152 236.906 145.234 235.344 143.188 230.256 143.371 225.142 148.092 221.969 147.859 220.969 144.5 218.515 143.552 217.758 139.406 216.047 137.891 214.188 138.075 213.401 139.951 210.587 141.101 205.473 140.345 200.344 143.312 198.483 142.04 195.625 143.891 193.516 142.375 192.344 140.281 195.484 137.547 196.609 133.875 196.297 132.391 196.734 130.516"
             ></Department>
             <Department
-              {...getDepProps("60")}
+              dep="60"
               points="288.418 73.801 287.178 75.284 286.421 77.16 287.935 77.16 287.178 79.066 286.421 82.849 287.541 84.725 287.541 87.963 289.447 87.963 288.69 89.264 287.359 91.745 286.814 93.439 288.503 94.952 289.048 99.309 290.022 100.822 288.69 101.185 286.996 100.247 286.693 102.486 286.814 102.335 287.541 104.029 288.69 105.905 293.799 106.299 297.403 105.905 299.854 104.029 302.875 105.905 304.392 107.055 306.663 106.48 308.721 105.542 312.715 107.63 316.861 110.081 318.193 111.413 320.463 109.9 322.339 111.02 323.489 111.958 325.178 111.776 326.328 110.263 328.96 111.776 332.198 110.444 334.074 111.019 335.95 109.506 337.098 108.931 337.458 109.203 337.854 106.631 336.492 105.088 334.192 103.545 333.223 105.088 332.649 105.269 332.463 102.394 334.192 102 333.796 99.337 331.496 98.944 332.646 97.007 335.914 96.251 337.064 91.651 338.789 90.864 336.489 89.139 337.276 87.414 337.637 81.664 336.852 77.216 332.858 77.61 330.165 77.247 325.173 78.579 320.941 82.603 317.491 81.453 314.042 81.09 311.354 78.397 306.542 77.035 300.036 77.61 298.307 76.278 294.828 76.278 292.348 77.246 291.202 76.459 291.202 74.371 290.806 73.796 288.415 73.796"
             ></Department>
             <Department
-              {...getDepProps("59")}
+              dep="59"
               points="316.077 0.055 310.024 2.899 300.583 4.412 300.069 4.503 303.67 10.222 306.182 16.577 310.388 16.759 312.117 18.091 310.029 19.059 310.029 21.359 312.687 24.234 317.499 25.414 323.461 25.777 324.792 23.659 326.73 23.659 327.668 25.596 325.187 27.139 325.583 30.982 329.032 31.92 332.482 32.314 333.451 34.795 335.176 35.582 334.782 36.914 332.301 38.064 332.089 39.214 334.027 41.544 335.359 42.482 334.996 43.632 333.453 44.782 334.027 45.176 336.506 45.358 338.624 46.72 338.442 48.445 337.293 49.201 337.081 52.288 335.356 52.47 333.627 52.288 331.721 53.438 332.871 53.832 334.777 53.832 336.502 54.195 337.077 55.164 335.957 57.312 337.471 58.825 338.989 59.219 340.503 58.251 342.591 58.251 343.165 59.401 343.922 59.219 346.193 57.887 348.463 59.219 351.489 57.131 352.82 57.131 354.334 58.463 357.359 56.375 358.661 56.557 359.807 57.495 363.989 57.888 364.347 59.583 366.435 57.707 367.554 57.707 368.312 60.158 371.916 61.096 372.945 60.4 372.637 60.4 372.455 58.524 376.238 56.254 375.663 52.653 372.059 51.715 372.999 50.747 372.999 48.114 375.844 46.026 375.087 44.513 369.035 39.792 358.443 40.367 357.323 42.243 355.992 42.243 356.174 35.616 353.153 32.045 350.882 32.408 349.551 30.895 345.768 32.589 344.467 31.288 341.804 30.894 341.047 28.443 340.865 20.878 339.171 20.122 338.959 18.972 337.84 18.972 337.446 16.702 334.995 16.914 330.271 18.427 328 21.237 325.73 21.237 324.211 19.361 323.637 17.273 321.76 15.185 319.099 15.185 317.983 13.127 317.983 9.89 319.284 7.832 318.528 4.988"
             ></Department>
             <Department
-              {...getDepProps("58")}
+              dep="58"
               points="336.382 193.453 335.444 194.936 333.658 194.936 331.262 194.392 328.605 195.179 328.786 197.085 331.057 199.718 331.057 203.319 329.367 205.77 329.912 208.04 333.332 210.703 333.514 213.336 335.42 217.33 335.026 221.87 337.114 223.928 336.539 228.861 336.352 231.312 337.29 233.006 335.838 237.636 335.958 239.058 339.196 240.571 341.466 243.052 343.528 241.902 345.429 240.782 345.825 242.476 348.637 242.476 349.394 240.963 351.088 241.719 351.663 244.17 353.176 243.807 356.78 238.874 358.656 240.206 358.959 240.781 361.925 238.935 363.135 239.086 364.073 241.386 365.858 241.114 367.214 239.752 368.97 239.752 370.331 237.997 371.692 237.725 371.963 236.757 374.929 236.908 375.08 236.212 373.719 235.001 373.719 233.791 375.625 232.701 375.625 231.884 373.84 230.795 373.569 228.768 373.715 226.861 372.504 226.044 373.566 224.561 374.534 224.017 375.2 222.383 374.262 221.839 373.173 220.205 374.534 218.298 376.834 216.937 379.679 216.937 379.679 214.789 380.496 214.517 380.496 213.7 379.285 211.794 376.561 211.249 375.623 209.615 376.019 208.132 376.29 206.074 375.81 204.985 372.785 207.043 371.545 207.588 369.91 206.498 370.06 203.775 368.276 203.775 366.794 204.743 366.4 203.502 367.212 202.14 366.244 200.93 365.154 202.564 365.3 203.774 362.733 203.653 358.648 199.84 355.531 199.719 355.531 197.661 353.351 196.299 352.923 194.544 352.258 194.423 352.258 197.812 351.047 198.084 349.265 197.539 347.389 198.749 346.296 199.021 345.207 198.083 344.117 198.627 341.666 196.993 339.911 196.993 338.67 196.176 338.942 194.663 337.581 193.453 336.382 193.453"
             ></Department>
             <Department
+              dep="57"
               d="M448.044,83.666 L445.199,83.847 L442.928,85.754 L442.353,86.692 L439.151,86.692 L438.031,85.542 L437.547,85.542 C437.622,87.169 437.698,88.387 437.698,88.387 L436.76,89.869 L439.181,95.044 L440.088,95.8 L441.208,98.766 L439.725,101.338 L441.39,103.002 L440.663,104.485 L440.663,106.543 L438.6,106.543 L438.418,108.934 L442.321,111.355 L442.134,114.321 L445.105,114.321 L446.584,116.348 L450.275,116.348 L453.997,121.008 L454.905,123.58 L457.326,123.58 L462.138,125.638 L466.409,128.21 L466.409,128.967 L469.193,129.148 L472.704,131.176 L476.032,131.176 L478.816,133.234 L481.962,135.443 L484.201,136.17 L487.59,132.78 L488.74,128.816 L487.222,127.303 L487.04,125.971 L489.279,121.795 L484.008,118.769 L480.987,121.039 L479.258,119.919 L480.019,118.225 L478.14,116.893 L475.661,115.743 L475.661,113.292 L478.14,112.535 L479.078,108.178 L480.772,106.302 L481.529,109.146 L483.617,110.084 L487.219,110.659 L489.095,112.535 L491.366,112.535 L493.272,111.203 L495.718,112.928 L497.049,112.928 L498.568,111.596 L498.568,108.751 L500.232,105.846 L499.836,105.967 L498.535,104.091 L494.752,101.821 L493.421,99.733 L488.882,100.127 L486.219,102.578 L479.804,102.76 L477.902,101.428 C477.775,101.196 476.84,99.553 476.026,99.098 C475.995,99.081 475.942,99.054 475.905,99.037 C475.878,99.026 475.838,99.013 475.813,99.007 C475.799,99.002 475.759,98.982 475.751,98.977 C475.749,98.977 475.725,98.977 475.721,98.977 C475.7,98.976 475.655,98.977 475.63,98.977 C474.744,98.977 473.027,97.982 472.818,97.857 L470.155,98.977 L469.968,101.247 L466.76,101.64 L464.859,98.039 L463.739,97.645 L463.739,95.012 L461.076,93.862 L460.894,89.323 L458.988,87.447 L455.024,85.541 L453.147,85.541 L452.573,85.935 L450.667,85.935 L448.044,83.666 Z"
-              {...getDepProps("57")}
             ></Department>
             <Department
-              {...getDepProps("56")}
+              dep="56"
               points="93.567 151.632 90.42 153.145 88.211 153.145 85.881 155.021 86.032 156.413 87.424 160.015 88.211 162.829 93.053 163.616 95.413 165.492 96.351 164.372 97.894 166.43 97.107 167.368 96.956 170.152 95.564 170.152 94.475 171.877 92.266 171.877 91.358 175.599 93.477 178.989 96.503 179.745 97.623 178.05 97.079 180.108 99.711 181.258 103.13 184.647 104.25 186.735 103.887 189.186 103.493 191.667 105.763 193.362 106.913 192.03 105.763 190.517 105.763 187.128 108.033 187.672 108.79 185.402 109.365 186.734 111.816 188.822 112.966 186.916 111.816 184.283 113.904 187.128 116.537 186.734 115.993 185.402 118.444 185.977 120.32 188.247 119.382 189.76 116.931 189.004 114.086 187.672 112.573 189.579 114.843 190.335 116.537 192.968 126.766 192.03 129.398 192.605 128.097 193.725 128.279 195.45 128.642 195.722 129.459 195.571 131.154 193.876 132.274 195.177 135.3 195.177 138.901 193.301 144.197 191.213 144.379 185.918 145.499 185.313 143.502 181.561 145.227 180.169 144.924 179.231 143.835 178.596 145.53 177.809 147.104 175.933 146.953 174.057 144.925 174.057 144.441 172.211 145.863 170.335 144.289 167.521 141.959 166.099 139.296 166.099 138.358 165.797 138.358 164.556 139.75 163.285 140.537 160.168 140.083 158.141 139.448 158.928 135.696 158.626 133.971 155.479 132.579 155.479 130.4 156.266 130.4 158.293 127.434 160.169 125.074 159.231 124.136 157.204 123.379 158.293 121.17 158.293 120.081 156.901 118.992 156.901 116.662 154.541 113.999 155.177 111.185 152.665 108.673 152.847 107.916 154.39 104.951 155.964 101.35 154.39 100.412 153.452 98.838 154.541 96.962 154.39 96.659 152.847 93.693 152.06"
             ></Department>
             <Department
-              {...getDepProps("55")}
+              dep="55"
               fillRule="nonzero"
               points="408.369 81.155 408.913 81.911 411.758 81.911 413.09 83.212 414.784 82.092 416.115 79.913 417.209 80.065 418.722 81.578 419.479 85.179 420.386 85.179 420.265 87.449 421.384 89.295 420.627 91.716 420.809 92.442 421.747 91.716 422.655 90.051 424.531 89.869 427.678 88.931 429.342 90.232 430.643 93.379 430.643 95.044 431.551 95.8 431.551 97.071 430.643 98.191 430.461 100.612 431.188 101.732 431.552 105.606 432.671 106.544 434.336 107.271 433.609 108.753 435.636 110.629 433.79 112.656 434.153 113.957 435.999 114.865 435.999 115.803 433.79 115.803 432.852 117.104 433.034 118.042 434.517 119.525 431.764 126.364 432.491 128.422 432.491 131.751 433.247 133.416 434.337 133.416 434.75 134.281 433.247 134.354 431.765 135.08 431.765 136.2 433.611 137.864 433.611 140.467 432.672 141.949 429.707 142.857 429.688 141.844 428.799 141.772 428.799 143.039 425.077 143.795 425.077 145.46 421.875 145.75 422.172 145.464 418.178 141.102 416.483 140.527 411.844 138.188 409.493 135.624 405.529 135.049 405.346 128.423 401.744 125.578 401.351 124.065 402.683 121.798 401.533 120.857 402.683 118.016 404.771 117.078 406.219 112.406 403.469 112.375 404.875 110.688 403.621 106.483 402.319 103.639 404.014 102.129 403.045 101.944 402.68 99.858 404.875 98.781 406.073 97.589 405.343 95.712 405.131 94.38 407.219 92.685 407.979 88.903 405.706 86.058"
             ></Department>
             <Department
+              dep="54"
               d="M427.225,81.759 L424.955,83.847 L421.747,84.029 L420.627,85.179 L420.386,85.179 L420.265,87.449 L421.384,89.295 L420.99,90.415 L420.627,91.716 L420.809,92.442 L421.747,91.716 L422.655,90.051 L424.531,89.869 L427.677,88.931 L429.341,90.232 L430.07,91.715 L430.639,93.379 L430.639,95.044 L431.547,95.8 L431.547,97.071 L430.639,98.191 L430.457,100.612 L431.184,101.732 L431.371,103.215 L431.553,105.606 L432.672,106.544 L434.337,107.271 L433.61,108.753 L435.632,110.629 L433.786,112.656 L434.149,113.957 L435.995,114.865 L435.995,115.803 L433.786,115.803 L432.848,117.104 L433.035,118.042 L434.514,119.525 L433.243,123.035 L431.761,126.364 L432.49,128.422 L432.49,131.751 L433.246,133.416 L434.336,133.416 L434.91,134.354 L433.246,134.354 L431.764,135.08 L431.764,136.2 L433.615,137.864 L433.615,140.467 L435.461,139.892 L438.245,140.074 L438.427,143.04 L439.547,143.433 L438.245,144.341 L438.058,145.279 L440.085,145.642 L441.387,147.306 L447.499,146.943 L448.801,144.522 L451.584,144.522 L452.704,143.614 L454.55,144.734 L456.215,144.159 L458.636,144.341 L460.663,143.614 L462.69,142.131 L463.804,143.251 L463.991,140.648 L465.47,140.073 L466.231,142.494 L468.44,142.676 L470.674,143.251 L471.582,143.432 L474.76,141.95 L476.424,140.83 L477.906,138.984 L480.872,137.864 L482.81,137.44 L481.716,136.381 L483.806,136.563 L484.202,136.17 L481.963,135.443 L478.817,133.234 L476.033,131.176 L472.705,131.176 L469.194,129.148 L466.41,128.967 L466.41,128.21 L462.143,125.638 L457.331,123.58 L454.91,123.58 L454.002,121.008 L450.28,116.348 L446.589,116.348 L445.11,114.321 L442.144,114.321 L442.326,111.355 L438.423,108.934 L438.61,106.543 L440.668,106.543 L440.668,104.485 L441.397,103.002 L439.732,101.338 L441.211,98.766 L440.091,95.8 L439.184,95.044 L436.763,89.869 L437.701,88.387 C437.701,88.387 437.625,87.169 437.55,85.542 L435.008,85.542 L431.589,81.759 L427.225,81.759 Z"
-              {...getDepProps("54")}
             ></Department>
             <Department
-              {...getDepProps("53")}
+              dep="53"
               points="216.064 137.894 214.188 138.076 213.401 139.952 210.587 141.102 205.473 140.346 200.359 143.372 198.483 142.041 195.639 143.947 193.551 142.434 192.219 140.164 189.405 139.014 187.499 140.527 184.291 140.133 184.109 145.066 185.047 145.822 185.047 149.605 183.14 151.512 183.14 154.538 183.715 156.414 183.715 161.892 185.047 163.224 185.047 168.338 181.839 167.945 180.507 168.701 178.056 174.753 177.087 178.172 176.785 178.838 180.204 179.534 183.018 179.534 186.135 181.561 187.406 180.018 193.004 181.712 195.213 182.983 200.206 182.801 202.385 180.622 205.835 181.711 207.681 181.711 207.711 181.741 208.316 180.319 208.316 174.691 207.681 173.753 208.014 172.815 210.677 172.664 211.918 171.575 212.069 170.637 210.677 167.187 211.131 165.795 214.248 165.492 213.945 164.705 214.732 162.829 214.096 161.407 214.883 159.864 217.697 157.504 217.364 155.023 217.848 150.968 218.937 149.092 222.478 147.942 221.934 147.912 220.966 144.492 218.515 143.554 217.758 139.408"
             ></Department>
             <Department
-              {...getDepProps("52")}
+              dep="52"
               points="406.438 131.418 401.994 132.174 401.419 136.955 397.395 139.83 395.307 139.104 394.308 141.192 394.126 144.067 397.576 148.879 400.844 150.029 401.601 154.053 401.813 160.953 400.667 162.678 397.396 162.89 395.283 164.796 396.826 165.765 397.462 168.186 399.732 169.457 403.126 173.058 407.48 178.929 405.03 181.561 406.724 182.711 406.909 185.919 409.931 185.737 410.688 187.25 412.959 187.431 414.648 186.493 417.674 190.276 418.069 191.214 422.215 190.276 423.002 188.37 423.184 188.4 423.184 186.493 425.818 185.343 428.842 186.493 431.687 185.343 433.382 185.343 433.956 181.56 434.894 180.44 433.205 180.258 433.018 177.988 435.652 177.413 435.838 175.9 438.5 175.9 438.5 173.449 440.589 172.693 440.014 171.18 440.589 170.817 438.86 169.456 436.802 170.212 436.802 166.248 431.506 163.615 432.626 158.501 434.315 157.351 433.77 155.656 431.291 155.293 430.747 152.812 428.476 152.812 425.813 149.241 422.792 149.03 421.486 147.154 423.18 145.46 419.186 141.102 417.496 140.527 412.957 138.257 410.506 135.624 406.542 135.049"
             ></Department>
             <Department
+              dep="51"
               d="M366.038,92.29 L363.767,93.259 L364.163,95.135 L360.168,95.135 L356.598,97.798 L356.598,102.881 L359.236,104.606 L359.992,106.301 L355.635,106.664 L355.09,108.389 L356.785,109.509 L356.028,110.659 L354.334,111.415 L354.697,112.716 L357.143,112.716 L358.116,114.048 L356.422,115.198 L354.909,119.162 L352.059,120.494 L351.09,122.552 L350.152,123.702 L350.364,124.822 L348.846,125.79 L348.452,128.423 L349.966,129.361 L350.722,132.206 L349.784,133.9 L350.358,135.232 L353.17,135.05 L353.17,135.988 L353.987,135.806 L357.255,139.256 L361.859,138.5 L367.245,134.838 L370.483,134.838 L373.752,132.538 L377.595,130.42 L380.47,130.632 L380.866,134.475 L384.315,139.65 L388.159,139.65 L393.545,138.5 L397.389,139.831 L401.413,136.956 L401.988,132.175 L406.434,131.419 L406.34,128.423 L402.738,125.578 L402.342,124.065 L403.673,121.795 L402.527,120.857 L403.673,118.013 L405.762,117.075 L407.279,112.354 L404.254,112.535 L405.948,110.659 L404.617,106.483 L403.315,103.639 L405.004,102.126 L404.036,101.944 L403.794,100.643 C403.755,100.649 403.523,100.673 403.523,100.673 L401.794,99.13 L399.857,101.067 L399.282,101.067 L398.525,100.099 L393.921,99.917 L393.137,101.067 L391.806,101.067 L390.657,98.555 L388.146,98.555 L387.571,99.13 L385.091,98.948 L382.003,96.648 L379.886,96.073 L379.128,94.923 L375.107,92.411 L370.503,92.29 L370.565,94.015 L369.233,94.378 L366.038,92.29 Z"
-              {...getDepProps("51")}
             ></Department>
             <Department
-              {...getDepProps("50")}
+              dep="50"
               points="155.119 71.168 154.362 73.074 158.357 76.282 158.357 80.458 156.844 82.334 157.782 83.272 158.357 83.666 157.963 87.267 159.295 90.293 163.653 95.195 164.591 99.552 165.529 100.884 165.529 107.693 167.799 112.233 167.799 117.528 165.348 122.43 167.981 129.239 172.157 130.177 172.52 132.084 170.462 133.022 166.891 133.022 167.466 135.413 168.586 139.014 171.824 141.859 173.337 142.222 174.85 140.164 176.545 139.952 178.603 137.501 180.509 139.014 182.779 139.014 184.292 139.77 184.292 140.133 187.5 140.527 189.406 139.014 192.22 140.164 192.281 140.285 195.519 137.531 196.608 133.93 196.305 132.357 196.759 130.511 194.883 128.635 190.041 125.488 186.47 125.186 182.869 120.646 185.835 119.557 187.076 117.227 185.533 115.805 186.925 114.564 188.317 115.653 190.829 114.11 192.403 111.598 193.008 109.117 191.919 106.908 192.554 106.121 191.132 103.791 192.705 101.763 191.464 100.19 189.891 102.217 187.712 100.976 184.111 97.375 183.96 95.832 185.049 94.743 184.716 92.655 182.93 93.109 182.749 88.57 177.816 82.729 179.329 78.947 181.417 78.947 179.541 73.833 171.401 73.439 167.044 76.465 162.142 73.257"
             ></Department>
             <Department
-              {...getDepProps("49")}
+              dep="49"
               points="176.786 178.837 176.362 179.684 176.09 179.623 176.302 180.319 176.302 182.498 178.329 183.739 179.57 187.037 181.446 189.367 178.965 191.576 179.116 194.36 180.054 195.147 183.473 194.693 186.287 197.326 186.922 201.23 185.046 202.501 178.963 202.35 175.059 203.106 172.88 204.226 172.88 206.072 174.907 208.584 176.934 208.584 177.085 211.247 177.872 212.034 176.934 214.213 175.361 214.667 175.361 216.089 174.574 216.876 175.512 217.33 175.996 217.179 177.388 217.179 178.962 219.206 178.901 221.173 179.264 221.234 180.838 222.172 185.68 222.172 188.252 224.351 191.853 222.808 197.724 224.14 200.357 222.415 200.357 219.782 205.471 219.389 210.192 218.451 215.306 218.27 215.881 219.601 217.213 220.721 218.514 218.451 222.115 214.669 223.567 214.669 225.716 206.922 228.742 203.321 228.53 199.175 230.437 196.694 230.437 195.574 229.499 194.424 230.104 193.183 227.532 192.184 219.574 187.342 212.22 185.163 209.406 185.012 209.406 183.136 207.681 181.714 205.835 181.714 202.385 180.625 200.206 182.804 195.213 182.986 193.004 181.715 187.406 180.021 186.135 181.564 183.018 179.537 180.204 179.537"
             ></Department>
             <Department
-              {...getDepProps("48")}
+              dep="48"
               points="349.97 337.223 344.674 339.129 343.16 342.52 339.771 340.249 337.108 348.57 334.379 354.835 338.343 359.707 338.072 363.459 340.734 365.336 340.734 369.875 341.546 376.29 344.754 377.652 344.483 379.769 349.022 378.982 350.626 379.769 349.628 380.646 355.468 384.55 360.461 383.521 361.314 382.312 360.618 380.617 362.676 380.103 365.581 382.826 370.544 383.342 372.783 379.922 372.783 377.016 374.143 375.473 372.933 375.141 372.933 371.206 370.205 368.301 372.444 367.938 373.625 366.915 374.563 365.103 373.625 364.527 374.169 360.534 370.962 357.055 369.63 350.094 364.818 343.951 361.307 344.798 360.546 342.165 358.645 342.165 358.282 344.436 353.38 345.948"
             ></Department>
             <Department
-              {...getDepProps("47")}
+              dep="47"
               points="228.411 347.996 226.414 348.267 224.719 348.783 222.298 349.297 221.632 351.174 222.479 353.049 224.537 353.231 224.204 354.926 222.661 355.44 220.936 357.502 218.031 361.95 216.669 364.341 217.184 366.913 218.031 368.641 216.155 369.824 214.279 369.824 213.916 371.729 215.126 372.754 215.308 374.629 213.916 375.991 210.89 376.472 211.344 377.015 212.04 381.13 215.792 380.797 218.364 381.795 220.603 381.978 221.632 383.189 219.574 385.251 219.06 389 219.514 390.18 220.24 388.666 222.479 390.573 225.505 387.547 226.837 389.424 230.045 388.848 233.464 388.487 234.977 385.828 240.636 385.278 243.481 388.092 244.449 387.154 246.325 386.578 245.569 383.945 248.413 383.189 252.014 382.433 251.227 380.162 252.377 378.83 253.315 375.226 251.227 372.955 252.559 368.628 255.404 370.322 259.58 369.564 257.674 365.388 256.161 359.728 259.944 359.544 260.7 357.549 257.735 356.31 256.373 354.765 254.315 354.765 251.924 356.974 250.048 356.826 249.534 355.615 250.895 354.765 250.895 353.556 249.352 351.863 247.99 351.679 247.143 352.705 243.36 352.705 243.027 351.347 241.817 351.162 239.759 352.854 236.007 352.854 234.282 354.397 231.377 353.371 231.044 349.801"
             ></Department>
             <Department
-              {...getDepProps("46")}
+              dep="46"
               fillRule="nonzero"
               points="273.2 335.714 274.501 334.594 274.531 334.406 274.952 334.318 278.977 332.441 283.777 334.582 287.499 339.908 289.647 339.637 292.855 336.428 293.896 338.229 296.062 336.188 298.998 336.973 300.088 343.934 303.208 346.333 300.874 351.956 303.208 353.021 302.236 356.225 303.779 356.436 303.5 357.375 302.236 358.857 298.816 358.857 298.12 359.736 296.426 359.553 293.339 362.973 292.315 362.641 290.106 364.516 290.771 367.936 292.497 370.508 291.146 371.833 291.083 371.604 287.024 373.414 284.266 372.749 282.576 374.807 279.671 374.443 279.125 377.771 277.771 377.792 276.958 376.104 276.403 375.141 274.86 375.29 271.955 378.377 268.354 378.227 267.839 375.991 267.839 374.297 267.506 373.938 266.271 375.229 264.708 375.229 263.058 373.419 261.515 373.419 260.486 371.357 261.792 369.833 261.697 368.816 259.276 368.91 257.646 365.333 256.159 359.741 259.875 359.583 260.883 357.087 262.583 355.417 262.79 352.548 266.875 352.208 269.542 348.438 268.449 348.01 268.333 346 271.438 345.583 271.688 343.75 273.229 342.729 274.854 339.729 273.2 337.803"
             ></Department>
             <Department
-              {...getDepProps("45")}
+              dep="45"
               points="304.396 151.511 301.945 153.782 296.407 154.266 296.195 154.629 296.74 155.718 294.985 157.201 295.379 160.742 293.9 161.014 292.927 164.524 291.292 166.309 288.992 166.309 286.813 167.519 285.603 166.157 284.241 165.613 282.759 166.581 282.759 168.064 281.669 168.881 279.914 167.792 278.825 168.609 278.946 169.819 278.431 170.939 278.825 171.06 280.429 171.06 280.308 172.149 279.224 174.025 279.224 174.993 280.308 174.993 281.246 176.082 281.518 177.171 279.37 179.319 280.431 181.922 280.431 183.132 282.216 184.342 284.666 184.614 286.27 185.976 286.694 188.306 288.449 190.333 290.356 189.788 291.168 188.033 294.164 188.427 294.83 189.123 296.736 189.123 297.674 188.034 304.609 188.306 306.243 190.727 308.149 191.544 309.783 193.057 311.811 192.785 312.355 192.089 313.566 192.089 315.079 193.996 318.196 194.147 319.008 195.206 321.192 198.202 322.13 199.019 323.341 198.898 323.492 196.174 324.036 195.902 324.853 196.023 326.21 197.809 327.148 198.202 329.085 197.415 328.782 197.082 328.601 195.176 332.378 194.056 331.808 191.968 331.233 188.942 328.782 185.553 328.208 183.465 331.99 183.465 334.653 181.558 335.016 178.744 333.322 176.837 338.254 172.691 338.254 169.272 335.775 166.639 334.837 163.613 331.417 160.375 326.696 163.038 326.334 161.525 324.245 161.343 323.671 162.856 321.794 163.219 316.686 163.037 314.592 164.369 312.903 162.856 315.924 160.768 315.742 157.56 313.471 156.41 311.568 153.566 306.484 153.203"
             ></Department>
             <Department
-              {...getDepProps("44")}
+              dep="44"
               points="168.01 176.083 167.828 178.171 163.682 178.746 161.775 180.622 161.019 182.71 149.308 183.285 144.375 185.918 144.193 191.213 138.897 193.301 135.296 195.177 132.27 195.177 131.15 193.876 129.455 195.571 128.638 195.722 129.606 196.388 126.005 199.596 126.762 200.353 127.518 201.866 125.612 204.529 127.7 205.649 131.301 206.405 131.664 204.892 133.752 207.555 137.171 207.555 139.622 204.892 142.83 204.892 139.441 206.586 139.623 208.493 140.379 210.187 138.291 212.275 136.021 212.275 136.414 215.119 140.56 214.363 145.493 218.902 145.251 219.205 148.822 221.535 149.609 223.411 152.878 224.833 155.39 225.136 156.479 228.101 160.382 229.039 163.015 229.523 164.437 227.949 163.015 225.922 162.41 221.534 163.651 220.142 165.376 220.142 166.163 221.867 165.679 224.984 166.768 225.771 169.582 224.833 170.52 221.716 169.279 220.445 172.396 220.445 173.485 218.417 174.574 218.568 177.086 220.928 178.902 221.17 178.963 219.203 177.389 217.176 175.997 217.176 175.513 217.327 174.575 216.873 175.362 216.086 175.362 214.664 176.935 214.21 177.873 212.031 177.086 211.244 176.935 208.581 174.908 208.581 172.881 206.069 172.881 204.223 175.06 203.103 178.964 202.347 185.047 202.498 186.923 201.227 186.288 197.323 183.474 194.69 180.055 195.144 179.117 194.357 178.966 191.573 181.447 189.364 179.571 187.034 178.33 183.736 176.303 182.495 176.303 180.316 176.091 179.62 172.55 178.743 170.856 176.473"
             ></Department>
             <Department
-              {...getDepProps("43")}
+              dep="43"
               fillRule="nonzero"
               points="337.591 317.369 338.377 316.25 341.797 315.736 343.673 314.191 345.688 313.969 345.731 312.829 347.098 312.201 347.453 312.922 348.303 312.466 349.665 313.346 351.984 313.141 354.114 312.314 355.475 313.495 357.031 313.516 358.229 315.734 360.062 316.469 361.118 314.295 362.495 315.4 365.422 315.5 366.436 313.603 368.154 315.219 370.545 315.915 371.219 315.234 375.113 316.428 379.289 313.008 381.347 312.824 387.404 315.457 387.041 319.061 390.067 318.877 392.458 321.633 391.394 326.27 388.733 327.389 387.614 329.295 388.813 332.995 386.101 333.077 385.846 336.582 382.868 336.859 381.35 341.58 376.811 341.58 371.913 345.182 369.13 349.438 364.888 343.957 361.337 344.722 360.535 342.19 358.647 342.163 358.277 344.357 353.416 345.887 349.993 337.256 349.516 337.369 349.556 335.8 348.518 334.233 346.515 332.977 347.498 331.466 346.715 330.683 347.589 330.373 348.791 329.299 346.757 329.076 345.787 327.715 345.386 324.146 344.221 322.907 343.347 319.885 339.04 319.846 338.212 317.463"
             ></Department>
             <Department
-              {...getDepProps("42")}
+              dep="42"
               points="368.126 265.353 364.706 265.928 364.162 268.016 365.493 270.649 365.856 281.422 361.317 281.633 361.13 283.328 364.156 285.779 362.461 287.472 361.887 291.65 364.158 294.671 366.216 299.573 371.154 302.806 372.843 309.039 369.241 312.429 369.816 314.7 375.112 316.427 379.288 313.007 381.346 312.823 387.404 315.456 387.041 319.06 390.067 318.876 392.458 321.632 394.213 321.148 397.484 320.544 398.331 316.822 402.84 314.159 403.264 308.198 403.415 308.104 401.206 307.772 399.144 308.587 397.485 307.559 399.543 305.111 398.938 303.262 392.553 302.236 387.199 297.301 387.199 295.637 388.434 294.61 388.434 293.186 387.012 292.337 388.222 290.489 388.222 287.826 385.775 285.555 385.775 283.284 384.11 281.65 384.11 279.802 383.296 276.714 384.536 275.474 384.748 271.785 388.649 271.785 389.675 270.544 388.434 268.486 388.434 266.64 387.411 265.823 386.682 269.696 384.594 269.696 383.081 271.209 381.931 270.059 375.879 269.121 373.608 270.453 372.091 270.453 371.728 269.121 368.884 268.546 368.702 265.52"
             ></Department>
             <Department
-              {...getDepProps("41")}
+              dep="41"
               points="254.798 162.919 253.527 164.553 252.014 167.185 253.89 168.91 253.708 172.874 252.952 175.506 251.045 175.506 251.045 179.107 248.412 182.527 245.568 183.647 244.236 185.341 244.811 186.763 253.98 187.066 254.767 187.762 254.646 188.428 253.284 189.124 253.556 191 254.766 191.817 256.673 190.455 257.217 190.455 259.396 192.089 261.303 192.24 261.575 197.112 263.058 201.046 263.058 207.855 264.147 209.489 267.809 209.489 270.654 213.695 271.713 214.603 271.743 214.512 273.649 214.24 275.283 212.878 278.551 212.727 278.823 213.272 279.489 211.789 282.334 210.821 284.785 210.972 285.995 210.548 287.63 212.061 290.505 213.695 292.926 213.544 292.805 211.093 293.894 209.883 294.989 209.762 295.927 210.821 299.74 210.427 302.161 209.065 301.889 208.127 301.223 207.31 301.344 205.524 303.129 202.286 305.429 201.318 305.429 199.29 305.974 198.08 304.495 197.535 303.521 195.477 301.075 194.811 300.954 193.994 303.404 191.967 306.159 190.605 304.616 188.305 297.681 188.033 296.743 189.122 294.837 189.122 294.171 188.426 291.175 188.032 290.363 189.787 288.457 190.332 286.702 188.305 286.275 185.975 284.671 184.613 282.22 184.341 280.439 183.131 280.439 181.921 279.377 179.318 281.523 177.17 281.252 176.081 280.314 174.992 279.223 174.992 279.223 174.024 280.314 172.148 280.433 171.059 278.829 171.059 276.378 170.363 274.078 172.542 271.475 172.814 267.269 171.725 266.059 169.001 264.425 168.063 263.487 165.884 260.763 165.884 259.674 165.067 261.308 164.129 261.429 163.312 259.674 163.312 256.678 163.856"
             ></Department>
             <Department
-              {...getDepProps("40")}
+              dep="40"
               points="174.728 358.709 168.736 361.795 167.193 361.856 163.834 379.835 159.477 396.481 158.145 402.923 157.025 407.462 154.211 412.307 159.658 415.03 164.621 414.699 167.526 412.128 169.917 414.004 168.556 415.364 169.585 415.727 170.432 414.516 174.729 414.848 177.12 412.973 181.569 413.155 183.294 412.276 186.018 412.276 186.714 413.155 188.923 412.461 190.981 414.004 193.734 413.155 197.335 411.944 197.85 412.794 203.327 412.978 206.051 411.435 204.719 409.105 206.051 405.501 207.927 403.05 207.352 399.842 208.865 398.329 206.626 394.546 208.502 392.275 210.59 391.879 212.466 392.636 215.129 390.365 216.067 393.21 217.005 394.541 219.093 393.967 218.911 391.519 219.516 390.187 219.06 389 219.575 385.247 221.633 383.189 220.604 381.978 218.365 381.791 215.793 380.793 212.041 381.126 211.345 377.015 208.954 374.111 207.925 373.777 208.288 377.197 208.288 378.226 205.02 378.372 201.601 377.186 200.905 372.889 198.514 370.322 196.82 370.173 196.638 368.628 194.762 367.418 191.675 366.57 192.522 365.541 192.522 364.517 191.493 363.665 190.313 362.642 186.863 363.156 184.836 364.851 183.444 365.033 181.235 363.49 177.967 364.851 176.424 363.822 177.816 362.128 177.967 359.889"
             ></Department>
             <Department
-              {...getDepProps("39")}
+              dep="39"
               points="427.801 206.768 427.892 207.283 427.135 207.283 426.741 210.309 426.197 212.397 426.554 213.728 425.44 215.241 424.29 216.754 424.29 218.267 421.083 219.78 420.145 221.293 420.508 222.413 420.695 223.381 419.388 224.138 419.177 226.589 420.695 226.77 422.42 229.615 424.508 229.615 425.628 230.947 426.959 230.947 426.563 232.46 422.209 233.004 422.603 234.517 424.121 235.485 424.121 237.361 423.547 237.936 424.695 239.449 426.386 242.475 425.629 245.683 423.723 247.014 423.935 249.647 426.206 250.404 426.962 251.554 425.056 253.067 420.093 253.672 422.423 254.58 426.206 259.876 428.657 260.996 428.657 263.659 431.501 263.266 435.103 259.302 438.129 260.815 438.129 263.085 443.607 263.085 451.474 254.43 451.141 254.248 451.504 250.284 454.349 246.894 452.447 246.138 452.629 244.988 450.329 244.776 450.179 243.414 451.661 241.932 451.265 240.449 450.452 238.543 453.843 237.453 455.083 235.698 455.354 233.519 452.63 230.916 450.724 230.401 446.518 229.04 446.518 225.227 446.247 222.382 442.853 222.654 437.557 220.868 438.373 218.992 439.584 216.117 440.008 214.241 438.646 212.455 436.199 210.821 435.928 208.793 435.958 207.643 434.294 207.583 433.356 208.521 430.908 208.521 429.269 207.159"
             ></Department>
             <Department
-              {...getDepProps("37")}
+              dep="37"
               points="244.812 186.765 245.175 187.612 239.698 189.125 238.397 191.031 236.097 189.518 237.247 193.301 235.159 193.301 231.376 190.638 229.5 194.421 230.438 195.571 230.438 196.691 228.531 199.172 228.743 203.318 225.717 206.919 223.568 214.666 223.81 214.666 224.566 217.692 228.167 218.448 228.167 220.718 232.707 222.05 232.707 225.651 232.525 227.921 238.184 227.921 243.117 226.771 242.723 224.683 244.236 223.745 245.749 225.833 247.081 226.408 248.201 230.947 251.439 234.336 251.802 236.999 254.737 239.904 256.402 239.753 258.46 237.847 259.943 230.131 260.881 227.407 261.577 224.018 264.573 222.929 266.6 223.322 267.962 224.563 269.445 222.112 270.807 220.871 270.807 219.389 272.562 219.268 272.986 217.482 271.473 215.455 271.715 214.608 270.656 213.7 267.811 209.494 264.149 209.494 263.06 207.86 263.06 201.051 261.577 197.117 261.305 192.245 259.398 192.094 257.219 190.46 256.675 190.46 254.768 191.822 253.558 191.005 253.286 189.129 254.648 188.433 254.769 187.767 253.982 187.071"
             ></Department>
             <Department
-              {...getDepProps("36")}
+              dep="36"
               points="285.997 210.55 284.787 210.974 282.336 210.823 279.491 211.791 278.825 213.274 278.553 212.729 275.285 212.88 273.651 214.242 271.745 214.514 271.473 215.452 272.986 217.479 272.562 219.265 270.807 219.386 270.807 220.868 269.445 222.109 267.962 224.56 266.6 223.319 264.573 222.926 261.577 224.015 260.881 227.404 259.943 230.128 258.46 237.844 256.402 239.75 254.737 239.901 255.221 240.385 255.221 243.986 254.464 246.256 258.065 249.101 259.941 250.977 262.967 251.371 264.117 255.335 266.175 256.455 265.812 259.3 264.299 259.572 264.844 259.693 269.595 260.056 271.289 258.362 274.315 261.388 278.098 257.031 279.792 258.181 282.063 257.969 282.819 258.363 286.243 258.545 287.363 255.519 296.835 256.669 300.437 257.607 302.076 257.425 302.167 255.912 304.073 254.005 303.677 252.371 302.587 250.223 302.982 249.406 303.132 246.955 303.798 246.138 303.949 245.594 302.31 244.111 301.765 242.204 299.198 240.843 299.198 239.209 301.074 238.12 301.074 237.031 299.317 235.548 298.77 234.61 300.13 233.914 300.01 232.704 302.158 230.919 302.037 230.132 300.281 230.132 299.192 228.891 299.192 228.225 300.13 226.591 300.13 225.381 297.951 222.506 298.375 219.934 297.135 218.996 294.567 219.117 292.541 219.934 289.814 219.541 288.455 218.451 288.184 217.634 290.368 215.849 290.514 213.7 287.639 212.066"
             ></Department>
             <Department
-              {...getDepProps("35")}
+              dep="35"
               points="151.911 127.726 150.096 129.784 151.155 135.503 152.729 136.108 153.667 137.984 152.729 141.434 151.791 150.028 148.825 150.633 147.403 149.241 146.011 151.269 143.348 152.207 141.32 156.564 140.079 158.138 140.533 160.165 139.746 163.282 138.354 164.553 138.354 165.794 139.292 166.096 141.955 166.096 144.285 167.518 145.859 170.332 144.437 172.208 144.921 174.054 146.949 174.054 147.1 175.93 145.526 177.806 143.831 178.593 144.92 179.228 145.223 180.166 143.498 181.558 145.495 185.31 149.308 183.283 161.019 182.708 161.775 180.62 163.682 178.744 167.828 178.169 168.01 176.081 170.855 176.475 172.549 178.745 176.362 179.683 177.088 178.17 178.057 174.751 180.508 168.699 181.84 167.943 185.048 168.336 185.048 163.222 183.716 161.89 183.716 156.412 183.141 154.536 183.141 151.51 185.048 149.603 185.048 145.82 184.11 145.064 184.292 139.768 182.779 139.012 180.509 139.012 178.603 137.499 176.545 139.95 174.85 140.162 173.337 142.22 171.824 141.857 168.586 139.012 167.466 135.411 166.891 133.02 157.601 133.02 154.182 130.932 156.452 127.906"
             ></Department>
             <Department
-              {...getDepProps("34")}
+              dep="34"
               fillRule="nonzero"
               points="323.37 416.961 320.434 414.268 321.221 411.605 322.825 409.214 325.488 411.061 332.448 408.67 333.537 406.521 338.348 406.521 338.077 399.047 339.681 399.047 342.617 400.65 345.552 400.65 345.038 397.443 346.642 396.081 349.85 395.566 349.85 394.568 352.573 395.506 352.573 397.049 353.784 397.745 354.812 396.867 356.174 396.867 356.87 398.41 358.928 398.41 359.956 394.809 361.651 394.809 364.223 391.572 367.309 391.902 367.824 396.352 369.005 397.744 370.911 396.716 374.148 398.408 375.359 400.468 380.988 403.889 383.046 408.699 383.046 411.24 379.444 413.299 377.225 415.331 374.511 415.205 357.867 427.885 356.173 429.973 350.877 430.155 350.121 431.85 343.311 433.938 343.19 434.966 340.769 434.875 339.945 432.225 337.531 431.97 335.291 428.701 333.566 428.883 331.539 427.672 331.175 424.769 330.148 425.282 330.589 427.184 328.272 427.34 328.09 430.76 324.519 431.971 322.794 428.369 320.403 429.912 318.345 428.369 317.316 425.978 319.041 423.92 318.225 421.711 322.28 420.444"
             ></Department>
             <Department
-              {...getDepProps("33")}
+              dep="33"
               points="176.12 301.363 173.094 305.902 172.156 321.79 169.705 337.857 167.98 350.326 167.798 353.564 169.13 349.207 171.763 345.787 175.546 349.207 175.939 350.326 177.059 351.839 172.338 352.051 171.582 350.901 169.706 351.663 169.312 354.508 167.224 357.357 167.224 361.681 167.194 361.867 168.737 361.805 174.729 358.719 177.967 359.901 177.816 362.138 176.424 363.832 177.967 364.859 181.235 363.5 183.444 365.043 184.836 364.859 186.863 363.166 190.313 362.65 191.493 363.675 192.522 364.527 192.522 365.551 191.675 366.58 194.762 367.428 196.638 368.64 196.82 370.185 198.514 370.332 200.905 372.901 201.601 377.201 205.02 378.382 208.288 378.234 208.288 377.204 207.925 373.785 208.954 374.117 210.891 376.478 213.917 375.994 215.309 374.633 215.127 372.758 213.917 371.731 214.28 369.826 216.156 369.826 218.032 368.643 217.185 366.917 216.67 364.345 218.032 361.953 220.937 357.506 222.662 355.445 224.205 354.929 224.538 353.236 222.48 353.051 221.633 351.176 222.299 349.299 224.72 348.783 226.415 348.266 228.412 347.995 228.291 347.902 228.14 344.15 230.016 342.79 227.625 341.247 225.234 344.15 219.394 344.333 218.879 342.939 217.185 342.089 218.547 340.361 218.547 338.486 217.851 337.458 217.851 336.433 219.576 335.404 220.09 332.349 221.119 329.593 220.09 328.047 218.214 328.047 217.215 326.866 216.246 328.891 214.37 327.559 211.707 328.891 209.437 328.531 205.11 324.175 202.447 323.992 201.691 317.936 196.758 317.364 196.576 314.516 195.638 315.454 190.039 315.454 190.311 316.665 191.643 322.144 192.006 327.623 191.068 329.137 190.13 324.598 187.467 314.217 177.814 305.502 178.026 301.537"
             ></Department>
             <Department
-              {...getDepProps("32")}
+              dep="32"
               points="240.636 385.277 234.977 385.822 233.464 388.484 230.045 388.847 226.837 389.418 225.505 387.541 222.479 390.563 220.24 388.66 218.908 391.506 219.09 393.957 217.002 394.531 216.064 393.2 215.126 390.35 212.463 392.621 210.587 391.865 208.499 392.26 206.623 394.531 208.862 398.312 207.349 399.827 207.924 403.035 206.048 405.486 204.716 409.087 206.229 411.75 211.01 412.083 212.886 411.084 215.791 411.084 215.64 414.322 217.849 415.533 220.088 415.865 221.117 418.619 221.632 419.134 221.299 420.828 220.603 421.524 222.661 422.88 222.812 424.424 225.384 424.424 226.08 423.733 228.289 423.733 228.985 425.276 231.376 425.276 232.072 426.669 236.702 426.669 240.152 428.454 240.455 428.03 242.331 427.002 245.75 422.553 252.104 423.073 254.828 424.949 255.705 423.739 257.067 419.654 258.762 415.717 262.363 414.171 264.088 413.477 263.574 411.934 261.849 411.75 261.183 410.057 259.791 410.057 257.582 407.82 257.733 406.277 254.828 403.373 254.314 401.525 252.438 402.34 251.924 401.313 253.134 399.436 251.742 398.078 251.742 395.684 250.562 394.473 246.779 394.289 246.779 392.08 248.837 390.537 248.837 388.81 251.076 387.783 249.866 387.117 248.504 387.783 245.751 387.783 244.903 387 244.449 387.15 243.481 388.088"
             ></Department>
             <Department
-              {...getDepProps("31")}
+              dep="31"
               points="276.586 395.506 273.832 396.535 272.985 397.896 271.956 396.711 270.08 396.528 269.717 398.254 268.537 398.738 270.231 399.615 269.051 401.492 264.754 402.702 262.878 400.463 261.184 400.463 259.792 401.316 254.829 401.316 254.315 401.527 254.829 403.373 257.734 406.279 257.583 407.822 259.792 410.06 261.184 410.06 261.85 411.75 263.575 411.937 264.089 413.48 262.364 414.171 258.763 415.72 257.068 419.654 255.706 423.739 254.829 424.949 252.105 423.073 245.751 422.553 242.332 427.002 240.456 428.03 239.246 429.721 237.884 431.783 237.884 432.993 235.675 433.841 233.95 436.746 233.768 438.289 236.854 440.528 237.369 442.737 236.34 444.462 238.398 443.772 239.427 444.649 240.274 448.215 237.883 449.611 238.065 451.487 236.855 452.668 235.161 451.82 233.254 451.82 233.496 461.988 241.183 462.503 241.576 453.213 244.209 453.609 248.113 455.88 249.717 454.401 249.021 453.373 249.021 449.953 250.715 447.891 255.194 446.68 255.86 444.138 254.831 443.448 254.68 440.177 258.614 437.786 260.823 437.635 263.062 438.301 264.938 440.03 266.148 440.176 267.177 438.487 266.299 437.273 264.756 436.757 264.605 436.242 265.785 434.547 269.9 434.185 270.415 433.187 269.719 431.976 267.328 430.25 267.842 429.222 269.052 428.374 270.232 428.707 272.138 431.279 272.289 433.004 275.558 433.337 276.224 431.794 274.681 429.917 275.71 430.25 279.492 431.976 281.55 431.643 283.065 431.007 282.706 430.613 285.127 429.523 285.642 426.585 288.609 425.253 287.796 423.377 290.487 420.956 291.816 423.377 295.809 422.046 296.568 421.925 296.63 420.685 296.63 418.265 294.57 418.626 292 417.933 289.761 415.206 288.913 413.996 284.465 412.12 283.438 410.577 284.798 410.059 284.798 408.184 283.438 406.639 281.711 403.887 281.561 401.493 281.046 401.159 278.988 398.738 278.141 396.195"
             ></Department>
             <Department
-              {...getDepProps("30")}
+              dep="30"
               points="374.571 365.094 373.633 366.909 372.452 367.937 370.213 368.295 372.942 371.2 372.942 375.138 374.152 375.47 372.791 377.013 372.791 379.915 370.552 383.334 365.589 382.819 362.684 380.091 360.626 380.608 361.322 382.3 360.469 383.512 355.476 384.538 349.636 380.637 348.516 381.634 348.516 384.327 346.37 384.841 346.915 386.989 349.578 387.534 352.514 387.534 353.33 391.257 349.851 392.861 349.851 394.554 352.579 395.492 352.579 397.035 353.79 397.729 354.818 396.852 356.175 396.852 356.871 398.395 358.929 398.395 359.957 394.791 361.652 394.791 364.219 391.557 367.305 391.889 367.825 396.337 369.006 397.729 370.912 396.703 374.149 398.395 375.36 400.455 380.989 403.875 383.047 408.687 383.047 411.228 379.445 413.288 377.115 415.378 380.019 415.591 380.019 419.371 384.342 419.159 386.672 419.28 388.366 415.343 394.51 411.59 393.178 409.713 394.782 405.444 400.38 406.231 401.742 396.061 409.217 391.794 409.217 389.372 403.588 383.771 403.588 379.474 400.38 374.148 393.692 370.396 393.172 373.333 390.484 373.604 389.697 370.666 387.004 371.184 386.489 374.938 384.341 374.151 379.802 371.184 377.656 372.274 377.656 366.92"
             ></Department>
             <Department
-              {...getDepProps("38")}
+              dep="38"
               points="424.653 285.355 423.413 286.808 422.172 290.496 419.693 291.734 417.03 289.255 415.179 289.255 414.967 291.948 417.66 294.219 413.545 299.541 408.191 300.782 404.076 302.232 406.739 304.894 407.374 306.132 403.259 308.194 402.835 314.155 402.714 314.217 403.864 316.609 407.163 317.638 409.433 316.822 412.096 314.974 415.611 317.849 418.486 317.849 420.544 320.724 419.696 322.784 420.12 325.687 419.092 328.562 419.515 329.589 420.941 329.163 424.241 330.19 428.569 331.428 430.415 330.19 431.227 328.766 431.868 328.766 432.05 344.167 433.079 345.195 435.767 345.195 438.218 346.646 440.067 348.099 441.945 348.282 443.155 349.308 446.575 349.704 446.812 349.131 445.299 348.372 445.299 346.101 450.021 346.101 451.538 344.586 450.6 342.894 453.051 340.989 454.78 341.745 457.226 339.657 461.553 340.415 463.071 338.721 466.673 338.907 466.673 334.761 465.159 334.003 464.401 331.554 460.433 331.158 459.862 330.22 460.62 326.255 461.952 325.107 460.86 323.627 458.799 322.387 457.558 323.627 457.985 321.961 457.985 320.328 456.323 318.663 457.171 314.762 459 313.74 458.788 311.078 454.885 307.174 453.432 307.174 452.409 308.596 449.958 305.328 448.505 305.51 447.265 308.203 448.117 309.837 447.477 310.473 445.842 309.233 441.091 308.209 438.852 304.093 438.852 302.429 436.582 299.982 436.34 298.92 428.778 289.691 427.927 287.027 426.293 285.576"
             ></Department>
             <Department
-              {...getDepProps("29")}
+              dep="29"
               points="78.709 121.099 76.651 123.369 74.381 122.431 70.205 122.825 69.45 124.7 66.999 125.275 66.605 123.187 62.278 123.762 62.278 125.094 59.252 125.276 57.92 124.338 56.407 125.095 56.013 127.365 50.929 127.547 48.266 130.755 50.536 132.45 47.51 134.931 48.448 136.626 47.692 140.772 50.718 141.165 51.868 140.015 52.443 140.772 59.615 139.834 64.336 136.445 60.19 140.409 60.553 142.285 64.336 140.591 63.58 143.254 67.756 143.435 67.574 144.555 63.035 144.374 59.434 143.436 55.077 141.348 52.444 144.374 55.833 145.524 55.651 150.607 56.589 149.851 58.677 146.643 62.641 148.913 64.548 149.306 65.304 152.332 64.154 154.39 61.703 154.208 59.433 154.208 55.65 154.783 49.205 155.146 47.904 156.871 49.78 157.991 51.868 157.809 53.562 159.322 56.013 159.14 60.008 163.679 60.946 168.581 59.614 171.244 63.578 172 67.935 171.789 68.873 170.094 67.179 167.824 68.873 168.58 70.598 168.398 73.624 170.093 75.5 169.73 75.5 166.492 76.256 169.73 78.707 173.694 84.003 174.057 84.214 172.937 85.515 174.813 88.753 175.388 91.204 175.388 91.355 175.599 92.263 171.877 94.472 171.877 95.561 170.152 96.953 170.152 97.104 167.368 97.891 166.43 96.348 164.372 95.41 165.492 93.05 163.616 88.208 162.829 87.421 160.015 86.029 156.413 85.878 155.021 88.208 153.145 90.417 153.145 93.564 151.632 92.898 149.241 93.836 148.303 93.534 142.826 92.112 140.95 92.112 138.619 94.321 137.984 91.961 137.197 91.81 135.351 92.899 134.564 94.473 132.99 92.748 131.901 92.899 129.874 90.569 129.723 91.658 124.639 88.541 122.824 83.245 123.006 83.245 126.788 81.732 126.788 81.369 125.094 79.099 125.457"
             ></Department>
             <Department
-              {...getDepProps("28")}
+              dep="28"
               points="278.674 118.224 277.524 119.162 277.524 122.188 273.741 124.064 273.741 126.908 272.621 128.24 267.87 128.24 265.631 127.302 258.792 130.873 256.159 130.873 253.435 133.476 254.071 133.899 254.253 137.137 258.247 140.526 258.61 147.516 256.734 149.786 253.89 150.361 251.045 152.055 251.439 153.568 252.74 155.293 252.74 159.772 253.133 159.833 255.766 161.709 254.798 162.919 256.674 163.857 259.67 163.313 261.425 163.313 261.304 164.13 259.67 165.068 260.759 165.885 263.483 165.885 264.421 168.064 266.055 169.002 267.265 171.726 271.471 172.815 274.074 172.543 276.374 170.364 278.432 170.939 278.947 169.819 278.826 168.609 279.915 167.792 281.67 168.881 282.76 168.064 282.76 166.581 284.242 165.613 285.599 166.157 286.814 167.519 288.993 166.309 291.293 166.309 292.928 164.524 293.897 161.014 295.379 160.742 294.985 157.201 296.74 155.718 296.195 154.629 296.407 154.266 295.887 154.327 295.493 149.243 295.097 148.668 294.739 146.217 290.773 145.461 289.044 143.373 288.499 139.197 286.228 138.834 285.832 136.746 283.199 134.87 281.868 131.632 283.199 129.362 281.868 127.849 281.868 125.973 282.625 123.885 281.109 122.372 280.537 120.102"
             ></Department>
             <Department
-              {...getDepProps("27")}
+              dep="27"
               points="243.662 87.69 242.028 88.084 239.849 90.414 235.612 90.958 236.489 97.979 238.396 98.554 238.396 99.31 236.489 100.43 238.759 103.638 239.909 108.177 239.697 110.083 238.002 111.203 238.184 112.535 240.09 113.292 240.272 114.986 238.759 119.344 240.272 121.25 245.568 121.432 246.688 124.065 248.594 125.971 251.227 126.909 252.014 129.754 250.683 131.63 253.436 133.476 256.16 130.873 258.793 130.873 265.632 127.302 267.871 128.24 272.622 128.24 273.742 126.908 273.742 124.064 277.525 122.188 277.525 119.162 278.554 118.314 278.463 117.467 279.431 116.499 277.737 116.136 277.737 114.623 276.769 113.11 277.525 112.172 282.82 110.659 284.152 108.389 285.298 104.213 286.694 102.488 286.997 100.249 288.691 101.187 290.023 100.824 289.049 99.311 288.504 94.954 286.814 93.441 286.84 93.32 283.849 92.987 278.826 90.415 274.469 90.536 272.169 92.563 270.929 96.104 266.057 96.649 261.427 98.011 263.061 99.1 262.395 100.734 260.216 100.734 259.006 98.131 257.644 98.131 255.465 96.497 258.461 95.559 253.831 92.29 249.231 92.562 248.142 90.686 245.539 90.958"
             ></Department>
             <Department
-              {...getDepProps("26")}
+              dep="26"
               points="412.096 314.98 409.433 316.827 407.163 317.644 404.892 316.947 404.892 324.664 406.345 330.414 407.373 333.713 409.008 337.828 407.984 340.491 406.133 344.002 405.709 346.059 403.863 346.455 404.886 351.205 404.886 355.291 402.407 358.193 401.802 361.882 402.83 363.519 402.226 364.972 402.226 366.421 403.038 366.607 403.643 367.455 403.428 371.45 405.943 371.087 407.063 372.418 406.518 376.2 407.274 377.138 410.086 374.29 412.567 374.104 413.112 372.59 409.904 372.195 409.36 368.804 411.418 365.597 414.081 365.386 416.714 367.865 414.081 371.435 414.837 372.765 418.62 373.161 421.071 371.073 419.739 373.918 420.133 376.189 424.101 376.551 429.398 376.945 430.155 379.396 433 381.484 435.451 381.667 437.144 380.155 438.114 377.308 439.627 378.64 440.746 379.972 441.503 372.951 439.627 372.589 438.479 370.139 432.067 368.412 431.28 365.205 433.58 363.691 431.674 362.359 432.07 360.666 435.091 360.85 437.724 362.182 439.817 359.731 437.724 358.039 437.911 355.768 439.057 352.164 443.021 351.768 446.229 350.467 446.562 349.709 443.142 349.314 441.932 348.287 440.056 348.103 438.21 346.65 435.759 345.2 433.071 345.2 432.042 344.173 431.855 328.77 431.219 328.77 430.407 330.197 428.556 331.434 424.229 330.197 420.93 329.168 419.507 329.593 419.084 328.566 420.112 325.691 419.684 322.789 420.536 320.728 418.474 317.853 415.6 317.853"
             ></Department>
             <Department
-              {...getDepProps("25")}
+              dep="25"
               points="471.587 189.519 471.405 189.73 469.221 189.73 468.107 191.273 466.679 192.09 466.679 193.997 463.139 194.421 461.237 193.059 458.666 193.452 456.215 195.359 454.459 198.355 452.704 198.9 451.887 200.927 449.708 201.321 446.713 203.772 442.778 203.651 441.568 204.589 440.474 204.589 436.662 207.706 434.934 207.645 434.906 208.795 435.177 210.823 437.628 212.457 438.99 214.243 438.566 216.119 437.355 218.994 436.539 220.87 441.835 222.656 445.224 222.384 445.496 225.229 445.496 229.042 449.702 230.403 451.608 230.918 454.332 233.521 454.061 235.7 452.824 237.455 449.429 238.545 450.246 240.451 450.64 241.934 449.161 243.416 449.308 244.778 451.611 244.99 451.636 244.808 463.348 233.824 462.99 224.745 467.136 222.687 469.986 221.355 472.619 218.904 472.831 215.303 475.463 213.971 481.516 206.981 480.578 204.711 482.666 203.773 485.117 200.747 483.786 199.415 479.247 200.353 479.06 199.596 483.241 194.784"
             ></Department>
             <Department
-              {...getDepProps("24")}
+              dep="24"
               points="242.542 296.764 240.848 299.79 238.185 300.152 238.003 304.691 229.107 310.744 228.925 317.371 225.506 320.762 223.63 322.491 219.847 322.095 217.759 325.699 217.215 326.881 218.214 328.063 220.09 328.063 221.119 329.608 220.09 332.363 219.576 335.421 217.851 336.447 217.851 337.473 218.547 338.502 218.547 340.378 217.185 342.106 218.879 342.955 219.394 344.349 225.234 344.166 227.625 341.263 230.016 342.806 228.14 344.166 228.291 347.919 231.045 349.822 231.378 353.393 234.283 354.419 236.008 352.876 239.76 352.876 241.818 351.184 243.028 351.369 243.361 352.727 247.144 352.727 247.991 351.7 249.353 351.883 250.896 353.578 250.896 354.787 249.535 355.636 250.049 356.847 251.925 356.996 254.316 354.787 256.374 354.787 257.736 356.332 260.701 357.57 260.883 357.088 262.608 355.396 262.79 352.549 266.936 352.189 269.599 348.406 268.449 348.011 268.267 345.95 271.475 345.554 271.687 343.677 273.2 342.709 274.894 339.683 273.2 337.804 273.2 335.715 274.501 334.595 272.806 331.935 272.988 327.789 268.842 327.972 267.148 326.823 268.661 324.948 266.573 323.256 268.086 321.352 266.573 320.594 266.573 317.961 270.356 314.541 268.268 312.846 267.148 310.002 263.153 309.428 261.821 308.49 264.666 307.158 263.728 305.856 259.582 305.283 258.614 301.502 252.562 300.929 251.23 302.834 249.929 303.195 248.204 300.924 248.991 298.836 248.023 296.959"
             ></Department>
             <Department
-              {...getDepProps("23")}
+              dep="23"
               fillRule="nonzero"
               points="278.098 257.031 279.792 258.181 282.062 257.969 282.818 258.363 286.242 258.545 287.362 255.519 296.834 256.669 300.436 257.607 306.609 256.844 308.937 260.058 308.937 262.328 310.203 263.938 311.547 262.406 313.087 264.991 315.175 265.748 317.263 270.65 317.619 276.518 319.133 278.395 319.859 282 317.619 283.689 317.047 286.375 311.385 289.562 311.734 291.188 315.679 295.189 313.785 296.656 310.06 296.656 308.848 298.505 306.578 298.505 304.688 296.594 303.906 295.438 299.159 295.417 298.156 294 296.679 294.01 295.844 296.016 292.781 296.674 292.531 297.016 291.568 296.469 291.172 294.062 292.531 294.016 292.578 292.984 290.116 291.327 287.453 290.118 287.828 287.656 287.034 287.639 286.005 289.091 282.707 289.275 280.878 289.023 280.895 287.265 283.076 285.381 281.678 283.52 278.209 284.47 278.806 276.712 277.169 275.073 275.505 270.959 273.091 270.304 272.63 268.084 274.896 263.568 274.828 260.782"
             ></Department>
             <Department
-              {...getDepProps("22")}
+              dep="22"
               points="106.913 115.44 105.188 116.772 100.86 117.316 99.892 118.648 96.866 116.378 92.902 119.041 94.415 121.099 91.782 124.7 91.661 124.639 90.572 129.723 92.902 129.874 92.751 131.901 94.476 132.99 92.902 134.564 91.813 135.351 91.964 137.197 94.324 137.984 92.115 138.619 92.115 140.95 93.537 142.826 93.839 148.303 92.901 149.241 93.688 152.055 96.654 152.842 96.957 154.385 98.833 154.536 100.407 153.447 101.345 154.385 104.946 155.959 107.911 154.385 108.668 152.842 111.18 152.66 113.994 155.172 116.657 154.536 118.987 156.896 120.076 156.896 121.165 158.288 123.374 158.288 124.131 157.199 125.069 159.226 127.429 160.164 130.395 158.288 130.395 156.261 132.574 155.474 133.966 155.474 135.691 158.621 139.443 158.923 141.319 156.563 143.347 152.206 146.01 151.268 147.402 149.24 148.824 150.632 151.79 150.027 152.728 141.433 153.666 137.983 152.728 136.107 151.154 135.502 150.095 129.783 148.885 131.145 145.284 130.751 144.921 132.839 142.651 133.021 142.469 130.388 140.563 129.813 139.231 131.326 139.231 127.543 136.961 129.237 133.572 128.662 132.422 130.932 125.432 134.715 125.432 136.621 123.919 136.621 123.919 133.202 119.955 131.326 120.318 127.906 116.747 125.274 116.747 122.066 114.084 121.491 114.266 118.465 112.208 118.283 112.39 116.195 108.607 116.195 108.032 118.071"
             ></Department>
             <Department
-              {...getDepProps("21")}
+              dep="21"
               points="391.215 167.761 390.821 170.212 388.37 171.544 382.137 171.726 382.409 173.087 382.555 174.449 380.654 175.811 380.654 178.111 381.048 178.535 382.561 178.535 383.499 180.139 382.955 182.862 380.776 184.496 380.776 186.13 381.466 186.674 381.32 187.068 379.959 187.612 379.565 189.64 377.381 194.27 375.902 196.449 375.902 198.628 376.446 199.566 375.63 200.655 373.724 201.593 373.845 203.5 375.63 204.589 376.296 206.072 376.025 208.13 375.629 209.613 376.567 211.247 379.291 211.792 380.502 213.698 380.502 214.515 379.686 214.787 379.686 216.754 379.836 216.815 383.497 220.598 387.309 220.477 390.702 223.049 393.153 224.835 393.273 227.135 395.841 227.68 398.025 229.435 403.745 227.408 407.679 226.167 409.436 225.895 409.98 225.108 411.886 225.229 413.368 226.167 415.547 225.622 417.726 224.14 419.365 224.322 419.39 224.14 420.697 223.383 420.51 222.415 420.148 221.295 421.086 219.782 424.293 218.269 424.293 216.756 425.443 215.243 426.557 213.73 426.2 212.399 426.744 210.311 427.138 207.285 427.896 207.285 427.713 206.165 426.956 205.408 426.744 201.625 425.05 201.444 424.687 198.418 422.599 197.45 423.357 196.512 424.505 196.149 426.956 193.305 426.744 191.792 425.226 188.766 422.992 188.373 422.205 190.279 418.059 191.217 417.665 190.279 414.639 186.496 412.949 187.434 410.678 187.253 409.921 185.74 406.896 185.922 406.714 182.714 405.019 181.564 407.47 178.932 403.116 173.061 399.723 169.46 396.701 167.765 391.222 167.765"
             ></Department>
             <Department
-              {...getDepProps("19")}
+              dep="19"
               points="296.68 294.01 295.868 296.069 292.782 296.674 291.571 298.731 290.119 298.731 288.057 298.127 286.604 300.577 284.37 300.79 282.918 303.452 281.066 303.452 279.614 304.905 275.922 304.48 274.681 306.539 273.259 306.357 270.778 309.443 268.508 308.595 267.448 310.741 268.265 312.829 270.353 314.521 266.57 317.943 266.57 320.577 268.083 321.333 266.57 323.239 268.658 324.931 267.145 326.806 268.839 327.954 272.985 327.77 272.803 331.916 274.377 334.395 274.952 334.305 278.977 332.428 283.789 334.574 287.511 339.9 289.659 339.629 292.867 336.419 293.926 338.297 296.074 336.178 299.01 336.965 299.101 337.602 301.552 335.874 300.795 333.999 299.649 332.85 301.162 330.975 302.675 330.975 303.647 327.951 304.406 326.436 303.831 323.016 306.282 319.624 309.701 317.535 309.883 312.24 311.578 313.178 313.483 315.266 315.36 315.266 316.506 313.936 315.568 311.665 316.506 308.275 315.931 305.068 314.785 303.736 314.785 301.074 316.511 298.262 316.117 295.598 315.692 295.174 313.787 296.654 310.061 296.654 308.849 298.503 306.579 298.503 304.733 296.654 303.88 295.415 299.159 295.415 298.13 293.991 296.679 293.991"
             ></Department>
             <Department
-              {...getDepProps("18")}
+              dep="18"
               points="306.151 190.608 303.397 191.97 300.946 193.997 301.067 194.814 303.513 195.48 304.487 197.538 305.966 198.083 305.421 199.293 305.421 201.321 303.121 202.289 301.336 205.527 301.215 207.313 301.881 208.13 302.153 209.068 299.732 210.43 295.919 210.824 294.981 209.765 293.886 209.886 292.797 211.096 292.918 213.547 290.497 213.698 290.349 215.847 288.167 217.632 288.438 218.449 289.795 219.539 292.524 219.932 294.546 219.115 297.118 218.994 298.358 219.932 297.932 222.504 300.113 225.379 300.113 226.589 299.175 228.223 299.175 228.889 300.266 230.13 302.024 230.13 302.145 230.917 299.997 232.702 300.117 233.912 298.756 234.608 299.301 235.546 301.059 237.029 301.059 238.118 299.182 239.207 299.182 240.841 301.752 242.202 302.297 244.109 303.934 245.592 303.785 246.136 303.119 246.953 302.968 249.404 302.574 250.221 303.664 252.369 304.06 254.003 302.157 255.91 302.066 257.423 306.509 256.878 307.416 254.941 309.867 252.127 314.588 251.159 317.069 251.915 319.702 249.857 319.518 248.344 318.58 247.588 318.58 244.562 323.664 239.448 325.57 241.536 327.26 239.811 328.777 239.63 331.41 236.24 335.766 236.603 335.828 237.632 337.28 233.002 336.342 231.308 336.529 228.857 337.104 223.924 335.011 221.866 335.407 217.326 333.501 213.332 333.319 210.699 329.899 208.036 329.354 205.766 331.044 203.315 331.044 199.714 329.076 197.414 327.138 198.201 326.2 197.808 324.838 196.022 324.026 195.901 323.482 196.173 323.331 198.897 322.12 199.018 321.182 198.201 319 195.205 318.188 194.146 315.071 193.995 313.555 192.088 312.343 192.088 311.798 192.784 309.775 193.056 308.138 191.543 306.235 190.726"
             ></Department>
             <Department
-              {...getDepProps("17")}
+              dep="17"
               points="172.336 288.288 174.999 288.288 174.424 287.35 174.242 284.718 176.512 284.901 177.45 280.723 175.755 278.272 176.512 277.153 178.782 277.516 178.964 275.428 177.269 274.671 176.301 272.038 173.668 271.463 172.912 269.768 170.831 269.864 170.355 268.98 170.643 267.864 173.487 267.501 173.85 265.595 176.876 263.144 175.484 261.873 181.264 258.363 183.927 258.181 185.042 261.203 188.643 259.69 191.487 261.808 191.578 264.592 193.636 266.831 193.636 268.072 196.723 270.342 200.627 270.554 203.29 274.034 207.618 273.64 211.128 276.086 213.791 277.326 214.214 279.388 215.697 280.81 215.636 280.836 214.002 286.8 213.579 290.701 212.157 291.305 210.099 291.12 209.675 290.096 207.405 290.096 206.195 291.12 203.502 291.12 201.232 292.331 202.897 292.971 203.108 297.295 203.925 297.479 202.472 299.143 204.137 300.17 206.376 302.231 207.617 304.29 209.07 305.712 208.646 307.773 207.829 308.587 209.07 309.828 209.07 311.28 206.589 313.339 207.618 314.155 209.071 314.579 209.071 314.974 207.013 315.79 207.195 316.818 208.436 317.454 212.34 316.818 214.004 318.664 215.245 320.329 219.27 323.054 217.757 325.687 216.244 328.894 214.368 327.564 211.705 328.894 209.435 328.535 205.108 324.18 202.445 323.996 201.689 317.942 196.756 317.368 196.574 314.521 195.636 315.459 190.037 315.459 188.615 308.923 182.563 302.871 180.475 302.475 180.293 300.6 177.63 299.083 173.666 295.848 171.396 296.061 171.002 291.885 170.687 288.453"
             ></Department>
             <Department
-              {...getDepProps("17")}
+              dep="17"
               points="170.254 287.084 169.883 284.141 165.526 280.934 165.92 274.307 168.553 277.514 171.761 277.91 172.154 282.451 173.092 283.207 173.411 284.186 172.545 287.214 171.093 287.545"
             ></Department>
             <Department
-              {...getDepProps("17")}
+              dep="17"
               points="170.024 267.911 169.129 266.169 165.346 265.806 164.953 263.899 162.683 262.386 159.475 263.535 161.563 266.561 164.196 266.561 166.859 268.256 168.947 269.95 169.897 268.7"
             ></Department>
             <Department
-              {...getDepProps("16")}
+              dep="16"
               points="239.093 271.98 237.459 273.22 237.67 275.908 236.641 276.72 233.736 275.056 230.861 276.085 227.563 276.298 225.082 273.271 224.295 273.423 220.996 275.056 218.515 275.697 218.515 277.12 217.093 278.784 217.486 279.808 215.64 280.837 214.006 286.798 213.583 290.701 212.161 291.305 210.103 291.118 209.679 290.094 207.409 290.094 206.199 291.118 203.506 291.118 201.236 292.329 202.901 292.969 203.112 297.295 203.929 297.479 202.476 299.143 204.141 300.17 206.38 302.229 207.621 304.288 209.074 305.71 208.65 307.769 207.833 308.585 209.074 309.826 209.074 311.278 206.593 313.337 207.622 314.153 209.075 314.577 209.075 314.97 207.017 315.788 207.199 316.815 208.44 317.452 212.344 316.815 214.008 318.662 215.249 320.327 219.274 323.05 219.849 322.081 223.632 322.476 225.508 320.75 228.927 317.36 229.109 310.733 238.005 304.68 238.187 300.141 240.85 299.778 242.544 296.753 243.603 296.784 244.239 293.909 245.752 293.727 246.69 291.063 248.596 290.489 250.866 288.975 250.503 284.25 252.197 283.68 255.586 282.918 255.405 281.229 254.83 278.596 252.197 278.381 250.684 276.69 251.622 274.995 251.683 273.452 249.988 272.604 246.72 272.604 245.267 274.239 241.757 275.69 240.547 274.45"
             ></Department>
             <Department
-              {...getDepProps("15")}
+              dep="15"
               fillRule="nonzero"
               points="321.945 311.287 323.489 314.525 326.516 314.469 327.756 313.346 329.814 314.373 330.843 314.523 333.415 318.458 336.138 318.641 336.834 317.279 338.212 317.463 339.031 319.797 343.347 319.885 344.221 322.907 345.4 324.113 345.732 327.714 346.757 329.076 348.672 329.328 347.609 330.286 346.781 330.656 347.391 331.438 346.515 332.977 348.518 334.233 349.516 335.766 349.516 337.369 344.674 339.124 343.156 342.422 339.771 340.244 337.108 348.562 334.359 354.766 332.724 352.775 332.172 347.438 330.359 347.359 329.516 342.078 325.781 342.531 325.266 339.156 324.701 338.312 319.345 342.878 319.312 347.344 318.014 347.417 317.469 351.688 316.138 352.232 315.078 354.624 309.178 354.624 308.663 353.837 305.97 353.837 304.366 356.5 302.297 356.188 303.234 353.094 300.938 351.922 303.208 346.333 300.088 343.934 299.1 337.604 301.5 335.891 300.794 334.001 299.703 332.844 301.161 330.977 302.625 330.984 303.646 327.952 304.404 326.438 303.829 323.018 306.28 319.626 309.699 317.537 309.906 312.359 311.531 313.203 313.438 315.266 315.297 315.297 316.504 313.938 315.566 311.667 316.135 309.593"
             ></Department>
             <Department
-              {...getDepProps("14")}
+              dep="14"
               points="235.583 90.808 231.165 91.595 223.963 95.771 215.853 98.979 209.407 95.378 193.913 93.108 190.312 91.232 184.714 92.654 185.047 94.742 183.958 95.831 184.109 97.374 187.71 100.975 189.889 102.216 191.462 100.189 192.703 101.762 191.13 103.79 192.552 106.12 191.917 106.907 193.006 109.116 192.401 111.597 190.827 114.109 188.315 115.652 186.923 114.563 185.531 115.804 187.074 117.226 185.833 119.556 182.867 120.645 186.468 125.185 190.039 125.487 192.793 127.273 196.606 126.123 199.42 122.855 203.323 123.944 206.742 121.584 208.8 120.827 210.979 123.006 214.58 122.37 217.697 124.095 221.601 122.854 225.172 120.191 227.532 117.528 229.106 117.225 229.56 119.253 230.801 118.95 230.952 117.528 234.553 116.923 235.794 117.679 239.637 116.832 240.272 114.986 240.09 113.292 238.184 112.535 238.002 111.203 239.697 110.083 239.909 108.177 238.759 103.638 236.489 100.43 238.396 99.31 238.396 98.554 236.489 97.979"
             ></Department>
             <Department
-              {...getDepProps("13")}
+              dep="13"
               points="407.042 392.055 401.746 395.08 400.384 405.248 394.786 404.461 393.182 408.729 394.514 410.606 388.37 414.358 386.676 418.296 392.668 418.568 400.626 419.138 402.14 420.656 399.295 420.656 397.419 423.864 405.528 425.553 411.969 424.433 408.549 421.201 410.82 419.324 414.422 420.838 416.116 424.442 426.919 424.625 429.733 423.477 430.305 425.206 427.282 427.839 431.458 428.022 430.702 429.927 429.552 431.258 438.817 431.258 443.356 432.773 443.805 433.377 443.992 429.624 445.35 428.081 447.079 427.052 446.894 426.027 445.535 424.667 444.174 424.667 443.296 423.639 444.84 422.246 444.84 421.726 443.15 420.879 443.15 419.517 446.903 419.704 447.781 419.009 444.51 415.922 444.697 412.32 442.635 410.625 444.329 407.207 448.444 404.453 445.355 402.391 443.143 404.12 437.999 405.303 433.883 404.786 426.529 401.728 422.081 401.876 418.329 400.181 416.933 398.28 414.034 395.042 407.195 392.139"
             ></Department>
             <Department
-              {...getDepProps("12")}
+              dep="12"
               points="324.701 338.312 319.345 342.877 319.345 347.416 318.014 347.416 317.469 351.687 316.138 352.232 315.078 354.624 309.178 354.624 308.663 353.837 305.97 353.837 304.366 356.5 303.791 356.439 303.458 357.499 302.248 358.859 298.828 358.859 298.132 359.738 296.438 359.556 293.351 362.974 292.328 362.642 290.119 364.519 290.784 367.937 292.51 370.505 291.178 371.901 291.482 375.805 294.21 375.987 292.667 379.074 293.877 382.16 295.42 381.283 296.115 382.826 298.325 380.617 301.594 380.434 304.832 382.826 310.49 383.854 312.548 387.456 315.453 388.817 316.996 392.752 316.809 394.295 318.686 397.745 318.686 399.621 322.105 404.069 325.376 405.762 327.434 405.244 328.463 403.887 330.006 404.245 333.546 406.516 333.576 406.516 335.145 406.516 338.352 406.516 338.081 400.914 338.081 399.039 339.685 399.039 342.623 400.643 345.558 400.643 345.044 397.436 346.648 396.076 349.851 395.559 349.851 392.869 353.33 391.265 352.514 387.541 349.578 387.541 346.918 387 346.373 384.852 348.519 384.338 348.519 381.645 350.636 379.768 349.032 378.983 344.493 379.768 344.764 377.651 341.556 376.294 340.744 369.879 340.744 365.34 338.082 363.463 338.353 359.711 332.725 352.776 332.18 347.422 330.334 347.422 329.517 342.068 325.765 342.613 325.25 339.133"
             ></Department>
             <Department
-              {...getDepProps("11")}
+              dep="11"
               points="304.91 420.713 304.639 424.193 301.16 423.103 297.408 423.103 297.679 421.772 295.802 422.043 291.808 423.374 290.477 420.953 287.789 423.374 288.601 425.25 285.636 426.582 285.121 429.52 282.7 430.612 284.846 433.003 284.301 434.607 293.682 439.143 294.469 445.588 294.469 449.067 295.014 453.606 290.202 453.606 288.87 455.483 295.014 460.567 298.493 458.69 302.76 463.774 302.125 463.835 302.941 464.319 310.628 460.567 308.752 457.813 308.601 454.575 326.545 454.575 326.213 452.154 330.328 449.945 335.109 453.697 337.588 454.843 337.44 449.426 337.654 443.197 335.383 443.379 333.476 440.567 334.995 438.086 338.203 441.112 341.047 438.841 342.954 436.964 343.194 434.967 340.769 434.876 339.922 432.152 337.531 431.966 335.292 428.697 333.565 428.879 331.539 427.668 331.176 424.765 330.152 425.283 330.668 427.341 328.276 427.341 328.09 430.761 324.519 431.972 322.794 428.37 320.403 429.913 318.341 428.37 317.314 425.979 319.043 423.917 318.226 421.708 318.014 421.77 312.385 421.77 306.515 420.708"
             ></Department>
             <Department
-              {...getDepProps("10")}
+              dep="10"
               points="377.598 130.419 373.755 132.537 370.486 134.837 367.248 134.837 361.862 138.499 357.258 139.255 353.99 135.805 353.173 135.987 353.173 136.562 350.54 137.682 350.358 139.952 349.027 141.677 348.268 145.46 347.841 148.637 349.027 149.605 351.297 149.605 356.199 154.719 355.624 159.833 359.044 162.465 360.95 161.133 363.971 164.735 364.334 168.154 366.422 170.966 367.36 173.054 374.922 173.448 378.704 171.542 379.824 173.63 381.155 173.812 382.124 171.724 388.357 171.542 390.806 170.21 391.202 167.759 396.681 167.759 397.441 168.183 396.802 165.762 395.257 164.793 397.375 162.887 400.646 162.675 401.792 160.95 401.58 154.05 400.823 150.026 397.552 148.876 394.105 144.064 394.287 141.189 395.286 139.101 393.53 138.496 388.144 139.646 384.297 139.646 380.85 134.471 380.454 130.628"
             ></Department>
             <Department
-              {...getDepProps("9")}
+              dep="9"
               points="269.051 428.369 267.841 429.217 267.327 430.245 269.718 431.971 270.414 433.181 269.899 434.18 265.784 434.542 264.604 436.237 264.755 436.752 266.298 437.266 267.176 438.477 266.147 440.171 264.937 440.02 263.061 438.295 260.822 437.629 258.613 437.78 254.679 440.171 254.83 443.442 255.859 444.132 255.193 446.674 250.714 447.885 249.02 449.947 249.02 453.367 249.716 454.391 248.112 455.873 249.141 456.447 254.981 457.567 257.462 457.567 260.67 461.743 268.81 461.347 272.018 466.46 274.862 465.314 283.184 466.46 283.789 469.94 289.659 468.578 291.805 466.702 295.013 465.886 295.799 464.282 302.759 463.765 298.492 458.68 295.013 460.558 288.869 455.474 290.201 453.595 295.013 453.595 294.468 449.058 294.468 445.579 293.681 439.134 284.3 434.594 284.845 432.99 283.062 431 281.543 431.636 279.485 431.969 275.703 430.243 274.674 429.91 276.217 431.787 275.551 433.33 272.282 432.997 272.131 431.268 270.225 428.7"
             ></Department>
             <Department
+              dep="8"
               d="M395.331,50.197 L393.424,53.011 L391.735,54.736 L391.735,56.43 L391.735,58.7 L389.464,60.213 L385.318,61.545 L383.047,62.483 L380.384,60.395 L376.782,60.395 L376.263,63.754 L377.957,66.962 L376.263,68.475 L376.081,71.32 L377.019,72.833 L376.262,75.103 L373.417,76.616 L373.417,79.067 L369.815,79.824 L369.24,81.337 L371.719,82.668 L370.963,85.301 L370.388,87.571 L370.509,92.292 L375.113,92.413 L379.134,94.925 L379.891,96.075 L382.009,96.65 L385.096,98.95 L387.577,99.132 L388.152,98.557 L390.663,98.557 L391.809,101.069 L393.139,101.069 L393.926,99.919 L398.53,100.101 L399.287,101.069 L399.862,101.069 L401.799,99.132 L403.528,100.675 C403.528,100.675 403.757,100.651 403.799,100.645 L403.678,99.858 L405.949,98.738 L407.069,97.588 L406.34,95.712 L406.128,94.38 L408.216,92.685 L408.973,88.902 L406.703,86.058 L407.459,84.726 L409.366,81.155 L409.91,81.911 L412.755,81.911 L414.087,83.212 L415.781,82.092 L417.112,79.913 L415.751,79.701 L415,75.919 L413.487,74.769 L408.191,74.194 L407.253,71.743 L405.527,70.623 L399.47,69.867 L399.112,65.51 L399.869,64.754 L399.869,63.059 L396.843,61.152 L397.418,59.064 L398.174,57.188 L396.843,56.038 L398.931,54.162 L398.931,50.743 L398.174,50.199 L395.331,50.199 L395.331,50.197 Z"
-              {...getDepProps("8")}
             ></Department>
             <Department
-              {...getDepProps("7")}
+              dep="7"
               points="402.714 314.225 398.326 316.829 397.474 320.551 394.205 321.155 392.45 321.639 392.511 321.731 391.393 326.27 388.733 327.389 387.614 329.295 388.37 331.746 388.915 333.077 386.101 333.077 385.889 336.648 382.868 336.859 381.349 341.58 376.81 341.58 371.913 345.182 369.185 349.509 369.638 350.084 370.97 357.044 374.177 360.523 373.633 364.517 377.657 366.939 377.657 372.293 379.803 371.204 384.342 374.17 386.49 374.957 387.005 371.204 389.698 370.687 390.485 373.625 393.173 373.354 393.693 370.418 399.837 373.867 400.835 371.84 403.439 371.445 403.651 367.451 403.046 366.603 402.234 366.417 402.234 364.965 402.838 363.515 401.81 361.878 402.415 358.189 404.894 355.287 404.894 351.201 403.871 346.451 405.717 346.055 406.141 343.998 407.992 340.486 409.016 337.824 407.381 333.709 406.353 330.41 404.902 324.66 404.902 316.943 403.877 316.611"
             ></Department>
             <Department
-              {...getDepProps("6")}
+              dep="6"
               points="487.168 366.06 484.596 367.418 484.414 369.814 482.537 371.357 481.509 373.777 480.818 378.226 482.876 381.283 483.058 383.855 484.244 384.213 487.689 387.45 490.926 390.35 489.568 391.227 487.506 389.836 484.934 391.227 484.244 392.08 481.515 392.08 481.848 393.436 481 394.465 481.696 396.194 479.457 397.553 480.304 399.615 482.543 399.762 484.753 400.789 485.115 404.391 487.174 406.453 488.354 406.453 490.593 407.305 490.926 409.207 489.57 410.393 490.593 411.416 490.593 412.626 492.711 414.811 492.862 414.624 493.044 410.296 496.826 411.059 498.158 409.33 500.064 409.726 500.246 403.855 504.602 403.494 508.381 400.075 511.775 400.075 511.957 397.984 515.377 395.925 513.471 391.57 516.32 389.121 515.746 386.273 519.922 384.943 521.041 380.797 520.497 377.952 519.528 376.258 518.767 373.779 515.955 373.993 507.059 377.2 504.212 377.2 499.314 373.233 494.378 371.901 491.533 371.901 491.533 368.481 487.568 366.034"
             ></Department>
             <Department
-              {...getDepProps("5")}
+              dep="5"
               fillRule="nonzero"
               points="459.618 326.256 460.95 325.107 464.35 324.751 465.47 325.87 466.062 328.484 469.435 328.137 469.831 325.293 471.525 324.535 475.078 324.703 477.337 329.361 480.149 330.3 480.922 335.906 485.843 338.439 488.477 338.043 490.564 338.8 491.139 344.279 493.772 345.611 493.719 347.234 491.713 347.12 489.875 349.078 487.478 348.939 485.417 352.025 482.514 353.753 481.5 355.75 478.915 355.962 478.922 357.734 478.219 358.797 477.188 361.359 471.048 361.285 468.146 359.742 466.239 361.105 462.67 360.92 461.859 362.078 462.67 362.13 463.188 365.363 462.311 365.727 459.072 363.668 459.078 362.453 457.196 360.915 456.168 360.915 456.156 362.578 454.624 362.973 451.356 364.85 449.294 368.298 448.779 369.993 449.989 370.325 450.141 373 448.962 373.079 447.084 371.354 446.141 371.562 446.575 373.079 449.375 376.281 447.604 377.01 446.242 376.161 442.792 376.161 439.888 378.855 440.503 372.953 438.627 372.591 437.479 370.141 431.066 368.414 430.328 365.219 432.531 363.688 430.673 362.361 431.069 360.668 434.062 360.844 436.656 362.156 438.781 359.734 436.724 358.04 436.906 355.703 438.056 352.165 441.953 351.797 445.219 350.5 445.781 349.156 444.298 348.374 444.312 346.188 448.984 346.109 450.5 344.609 449.598 342.896 452.016 340.984 453.766 341.734 456.203 339.703 460.516 340.422 462.062 338.719 465.609 338.891 465.641 334.766 464.157 334.004 463.406 331.562 459.432 331.159 458.86 330.221"
             ></Department>
             <Department
-              {...getDepProps("4")}
+              dep="4"
               points="487.504 348.934 485.446 352.022 482.541 353.751 481.513 355.813 478.946 355.96 478.946 357.835 478.25 358.859 477.222 361.431 471.078 361.285 468.174 359.742 466.267 361.099 462.702 360.917 461.822 362.127 462.702 362.127 463.217 365.364 462.339 365.727 459.101 363.665 459.101 362.455 457.225 360.912 456.2 360.912 456.2 362.641 454.654 362.973 451.385 364.85 449.326 368.294 448.807 369.989 450.017 370.321 450.204 373.075 448.994 373.075 447.115 371.346 446.089 371.533 446.604 373.076 449.508 376.309 447.632 377.004 446.271 376.157 442.823 376.157 439.921 378.881 439.891 378.854 439.77 379.971 438.652 378.639 437.138 377.307 436.169 380.158 434.475 381.67 433.747 381.608 433.773 381.792 433.924 384.364 435.983 384.723 435.983 387.118 434.106 387.968 434.439 390.724 437.859 392.414 437.859 394.29 436.315 396.019 436.315 397.562 440.614 397.562 444.551 402.009 444.521 402.101 449.514 400.466 455.688 404.579 459.108 402.009 459.959 400.282 464.771 398.739 467.831 400.796 472.277 397.562 477.423 397.746 478.781 396.353 481.656 396.232 481.718 396.202 481.022 394.476 481.873 393.448 481.54 392.087 484.266 392.087 484.96 391.239 487.53 389.845 489.592 391.239 490.95 390.362 487.714 387.457 484.266 384.221 483.083 383.858 482.899 381.29 480.84 378.229 481.533 373.785 482.559 371.36 484.436 369.817 484.622 367.426 487.191 366.065 487.587 365.917 487.587 362.256 490.249 361.896 488.733 360.566 486.83 359.991 485.892 357.541 486.65 355.848 490.041 352.245 489.497 349.582 489.977 349.066"
             ></Department>
             <Department
-              {...getDepProps("3")}
+              dep="3"
               points="331.419 236.242 328.786 239.632 327.273 239.813 325.579 241.538 323.673 239.45 318.589 244.564 318.589 247.59 319.527 248.346 319.709 249.859 317.076 251.917 314.595 251.161 309.874 252.129 307.423 254.943 306.516 256.88 306.667 256.85 308.937 260.058 308.937 262.328 310.243 264.022 311.575 262.328 313.088 264.991 315.176 265.748 317.264 270.65 317.355 272.07 320.23 274.31 321.775 273.615 322.988 270.71 324.169 270.377 324.169 268.834 326.228 268.652 326.414 269.681 328.981 266.776 331.886 266.776 332.404 267.805 331.045 269.681 333.068 271.921 333.43 273.281 338.211 276.036 344.021 276.884 345.75 276.702 348.321 277.216 350.53 275.856 352.257 276.708 352.589 279.101 354.829 279.617 357.728 279.468 358.581 281.526 361.244 282.588 361.335 281.65 365.874 281.439 365.511 270.665 364.18 268.032 364.724 265.944 367.871 265.4 367.962 265.188 371.926 262.162 372.109 254.778 370.78 252.902 367.756 252.902 366.637 251.389 363.429 251.389 362.491 250.239 362.491 247.394 358.679 240.222 356.803 238.89 353.2 243.823 351.683 244.186 351.109 241.735 349.419 240.979 348.661 242.492 345.848 242.492 345.453 240.798 343.548 241.918 341.489 243.068 339.218 240.587 335.98 239.074 335.799 236.623"
             ></Department>
             <Department
-              {...getDepProps("2")}
+              dep="2"
               points="357.353 56.37 354.332 58.458 352.818 57.126 351.487 57.126 348.461 59.214 346.191 57.882 343.92 59.214 343.163 59.396 342.589 58.246 340.5 58.246 338.986 59.214 338.924 59.214 339.378 61.877 336.899 64.57 336.899 67.051 335.17 68.988 335.566 71.288 336.504 75.131 337.654 81.667 337.291 87.417 336.504 89.142 338.804 90.867 337.079 91.654 335.929 96.254 332.661 97.01 331.511 98.947 333.811 99.34 334.207 102.003 332.478 102.397 332.665 105.272 333.24 105.091 334.208 103.548 336.508 105.091 337.87 106.634 337.474 109.206 339.203 110.447 339.778 114.623 344.862 119.525 346.551 120.1 347.52 122.37 350.702 123.036 351.096 122.552 352.067 120.494 354.915 119.162 356.428 115.198 358.122 114.048 357.151 112.716 354.7 112.716 354.337 111.415 356.031 110.659 356.788 109.509 355.093 108.389 355.638 106.664 359.995 106.301 359.239 104.606 356.601 102.881 356.601 97.798 360.171 95.135 364.166 95.135 363.77 93.259 366.041 92.29 369.248 94.378 370.58 94.015 370.396 87.569 370.968 85.299 371.724 82.666 369.245 81.335 369.82 79.822 373.422 79.065 373.422 76.614 376.267 75.101 377.028 72.831 376.09 71.318 376.274 68.473 377.966 66.96 376.274 63.752 376.791 60.393 372.948 60.393 371.919 61.089 368.315 60.151 367.556 57.7 366.44 57.7 364.349 59.576 363.991 57.881 359.81 57.488 358.664 56.55"
             ></Department>
             <Department
-              {...getDepProps("1")}
+              dep="1"
               points="410.491 250.404 408.403 250.615 407.072 253.066 403.501 267.047 402.986 268.197 402.623 272.614 401.595 274.042 401.595 280.427 400.991 281.879 404.679 284.118 406.131 284.33 408.402 286.39 408.796 289.693 411.277 288.875 415.12 289.994 415.181 289.265 417.032 289.265 419.694 291.744 422.173 290.506 423.414 286.818 424.654 285.365 426.289 285.576 427.923 287.029 428.771 289.693 436.371 298.954 438.641 297.318 439.037 293.804 441.516 293.38 441.516 287.2 442.757 286.172 443.15 280.634 443.634 281.028 443.573 279.006 442.545 277.13 442.969 271.806 444.815 272.833 445.843 270.956 447.69 270.351 449.596 268.838 447.538 268.838 447.538 265.237 449.808 263.905 453.201 263.542 453.414 261.636 452.265 260.88 455.109 257.279 454.716 256.159 451.478 254.434 443.61 263.089 438.133 263.089 438.133 260.819 435.107 259.306 431.505 263.27 428.66 263.663 428.66 261 426.21 259.88 422.427 254.584 419.008 253.252 417.888 250.801 415.982 250.408 414.106 251.739 412.592 252.133"
             ></Department>
             <Department
-              {...getDepProps("2B")}
+              dep="2B"
               points="549.137 436.703 546.293 438.609 546.686 440.485 548.205 442.386 546.51 443.693 547.267 445.206 546.121 446.538 546.121 448.232 548.027 449.957 548.027 452.59 546.877 455.041 545.576 455.616 544.058 453.528 541.395 453.74 540.825 453.344 538.554 453.344 536.496 455.25 535.709 458.458 530.807 459.396 527.029 462.603 526.268 464.692 524.392 464.505 523.423 463.359 522.878 466.591 521.547 467.137 521.153 470.162 521.723 471.493 519.635 473.007 519.064 474.525 521.178 474.919 521.541 475.947 525.293 475.947 526.322 476.613 529.076 476.093 530.256 476.789 529.742 477.641 531.618 480.209 535.219 480.209 536.098 482.78 538.639 482.78 538.493 484.324 540.551 486.896 541.58 487.411 542.939 488.258 542.939 491.529 543.965 492.376 546.027 492.376 546.542 493.229 546.874 497.859 548.085 498.707 547.57 499.403 547.901 502.126 551.837 501.975 555.136 499.796 555.015 494.198 559.554 487.782 559.554 477.191 557.678 473.587 557.105 462.239 555.775 460.151 553.326 458.275 552.93 451.254 554.079 448.046 552.564 442.937 551.626 438.791 550.84 437.641"
             ></Department>
             <Department
+              dep="2A"
               d="M517.545,474.529 L517.545,476.617 L519.451,477.949 L522.658,479.825 L522.87,481.338 L520.964,481.913 L517.938,482.488 L517.938,483.789 L519.058,484.939 L519.27,488.716 L523.416,490.047 L524.93,490.41 L526.261,492.498 L525.323,493.83 L523.805,494.375 L522.659,496.463 L521.54,497.794 L522.085,501.189 L524.93,501.002 L525.686,501.577 L528.349,500.245 L529.106,501.002 L527.775,503.849 L529.106,505.178 L526.836,506.872 L525.322,510.263 L529.468,511.236 L535.338,511.78 L532.887,514.625 C532.887,514.625 531.74,514.181 531.223,514.413 C531.207,514.421 531.176,514.434 531.162,514.443 C531.158,514.447 531.136,514.465 531.132,514.468 C531.128,514.476 531.106,514.494 531.102,514.498 C531.098,514.502 531.08,514.524 531.076,514.528 C531.065,514.537 531.053,514.549 531.047,514.558 C531.044,514.563 531.05,514.579 531.047,514.584 C531.042,514.599 531.021,514.637 531.017,514.646 C531.015,514.652 531.019,514.667 531.017,514.675 C531.014,514.689 530.991,514.724 530.987,514.737 C530.986,514.751 530.987,514.783 530.987,514.799 C530.987,515.742 529.686,518.035 529.686,518.035 L531.562,520.097 L534.981,522.185 L541.397,523.874 L543.273,524.636 L545.002,525.393 L543.854,527.481 L546.878,527.296 L547.45,528.625 L550.475,528.625 L551.232,525.024 L549.327,524.628 L551.988,521.816 L551.05,520.848 L551.232,519.153 L554.651,517.277 L554.838,515.189 L552.567,515.005 L551.054,516.308 L551.054,514.432 L554.077,514.245 L555.015,511.975 L555.772,505.348 L555.197,502.5 L555.136,499.774 L551.837,501.953 L547.899,502.104 L547.566,499.381 L548.086,498.685 L546.875,497.835 L546.543,493.204 L546.023,492.354 L543.965,492.354 L542.936,491.507 L542.936,488.236 L541.579,487.389 L540.551,486.872 L538.49,484.302 L538.639,482.758 L536.097,482.758 L535.219,480.187 L531.618,480.187 L529.741,477.617 L530.256,476.767 L529.075,476.071 L526.317,476.59 L525.293,475.925 L521.541,475.925 L521.178,474.897 L519.06,474.501 L517.547,474.501 L517.545,474.529 Z"
-              {...getDepProps("2A")}
             ></Department>
           </g>
         </g>
